♻️ refactor(web): port site.py into web.ts site and split web.ts into lib modules
One runtime, one entry point: the web skill is a single bun program, so
site.py's four commands become `web.ts site list|info|run|add` with
identical behavior and output (verified byte-for-byte against the Python
runner before deletion, live against real Lightpanda). web.ts is now the
CLI entry only; search/fetch/site live in scripts/lib/ modules.
Skipped a lib/lightpanda.ts: fetch and site spawn the binary differently
(spawnSync render vs piped run), so there is nothing shared to extract.
