+++ resources/builtin_manifest_gen.go
@@ -2,4 +2,4 @@
 
 package resources
 
-const builtinManifestJSON = "{\"revision\":\"59ca70b1988440309b5c92575881f64eea75d195ceb7a2c87575d6b9f172761c\",\"skills\":[{\"ref\":\"builtin:email\",\"api_id\":\"builtin-email\",\"name\":\"email\",\"description\":\"Standard IMAP/SMTP email client for Gmail, Outlook, and self-hosted accounts.\\nUse when user mentions email, inbox, send email, check mail, read message,\\ncompose, reply, set up email, add email account, configure email\\n(non-Lark/Feishu accounts).\\n\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"author\":\"CherryHQ/stella\",\"owner_plugin\":\"system/email\",\"version\":\"1.0\"}},\"root\":\"system/email\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"4df3c9a5f9f858c23745f66b95de788beb56994402a4164fc330817e09c021c4\",\"size\":3192,\"mode\":420}],\"digest\":\"bbd0a4528f9a59eb876ccc178571e07a0906c087d53e9bac2a4bbe05ce6c25cb\"},{\"ref\":\"builtin:html-artifact\",\"api_id\":\"builtin-html-artifact\",\"name\":\"html-artifact\",\"description\":\"Create polished single-file HTML artifacts for plans, reports, explainers, code reviews, lightweight dashboards, prototypes, and one-off editing tools. Use when the user asks for a standalone HTML file, single HTML, HTML artifact, interactive report, shareable local artifact, Alpine.js artifact, custom CSS artifact, copy/export UI, or a browser-openable deliverable instead of Markdown.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{},\"root\":\"system/html-artifact\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"03d360305bbcd78a6e50e934e06981dd54ab2d6cb66594884d5612e562f8e1fc\",\"size\":3284,\"mode\":420}],\"digest\":\"45d1e904378dee115664002fb8c3bfa96c84a5ad39c29e7b95ad08d89419ecc0\"},{\"ref\":\"builtin:lark-cli\",\"api_id\":\"builtin-lark-cli\",\"name\":\"lark-cli\",\"description\":\"Lark/Feishu CLI skills for Stella sessions. Covers workspace operations — calendar,\\ndocs, tasks, mail, and messenger — via the Stella-managed lark-cli tool. Stella installs\\nthe binary and injects a managed per-user OAuth token; no native CLI bootstrap is\\nrequired. Always\\nread lark-shared first for identity selection and permission-denied handling.\\n\",\"tags\":[\"lark\",\"feishu\",\"workspace\"],\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"generated\":true,\"owner_plugin\":\"tool/lark-cli\",\"source_ref\":\"2e4cfb49219d\",\"source_repo\":\"larksuite/cli\"}},\"root\":\"system/lark-cli\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"54ea27ac48fdb76072c2e2d1b6066d26b3add41f8caf5ddfa4addbe5ad1409a0\",\"size\":25555,\"mode\":420},{\"path\":\"references/lark-approval.md\",\"digest\":\"c03b4f65977087d63d7720768592a2f834ac6a21767a4d27572205fd3ae3e591\",\"size\":1639,\"mode\":420},{\"path\":\"references/lark-attendance.md\",\"digest\":\"776b012dd3d6bc50e510cc39ef416e94b636523f88af26f24a12ec7eaa00e4a1\",\"size\":1720,\"mode\":420},{\"path\":\"references/lark-base.md\",\"digest\":\"c95ea3c9b14d9f910cd711286bcd13651fae3dd60c58ce49e2bc85dca9072008\",\"size\":55109,\"mode\":420},{\"path\":\"references/lark-base/dashboard-block-data-config.md\",\"digest\":\"0f55f98760e76ef8c0eecfa9cfb0198c9a3f56be4387b910ac8c4640e0bdce00\",\"size\":15182,\"mode\":420},{\"path\":\"references/lark-base/examples.md\",\"digest\":\"49c4b3f1ab3770a0a2f6e82d9f6885f9310d40e0c39ef3dfe19c0e07172f2461\",\"size\":3637,\"mode\":420},{\"path\":\"references/lark-base/formula-field-guide.md\",\"digest\":\"72b14241bbb55c7cf50481c689a45fcd4462c13bf651031b689f740bdaecbf7d\",\"size\":56980,\"mode\":420},{\"path\":\"references/lark-base/lark-base-advperm-disable.md\",\"digest\":\"88f6f51163d6bae21bf5b5d681fbe0194cf73eed14263f03bdb798b24914be5b\",\"size\":2485,\"mode\":420},{\"path\":\"references/lark-base/lark-base-advperm-enable.md\",\"digest\":\"d73c18457807ae7f08ea5762f60f6659f2fa7328103ccb78a07d90f96d95b56d\",\"size\":2307,\"mode\":420},{\"path\":\"references/lark-base/lark-base-base-copy.md\",\"digest\":\"5cc2fe1478a299402345411d5d4c1cdfa6fa46fec4e5bd12d34178bd6f9b1467\",\"size\":2938,\"mode\":420},{\"path\":\"references/lark-base/lark-base-base-create.md\",\"digest\":\"89ff76f094acf6a75793726f3025ccbd8850cc3a36a7cc5e2b61b5bcc1a904b4\",\"size\":2606,\"mode\":420},{\"path\":\"references/lark-base/lark-base-base-get.md\",\"digest\":\"7bd335528ac4f0a312bd13667fd7836ea2b25accc3b5b65f16fdd5b0c6d44e3f\",\"size\":1025,\"mode\":420},{\"path\":\"references/lark-base/lark-base-cell-value.md\",\"digest\":\"7a3f9bfe433de84fa040b9e4a72974494d84490ee49eff35b49a290996ae0747\",\"size\":4152,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-arrange.md\",\"digest\":\"2edf273452d14af35abecc0b75d3820f6ad9231c0abe74de91f903cd1b9c6d97\",\"size\":2816,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-create.md\",\"digest\":\"4892f407b5ddb1967c1aa2d9dbe2f577a9d1ad898c3a1d416b092d02b06823a1\",\"size\":5374,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-delete.md\",\"digest\":\"685539ea10fd53862aedf0cf84a03771f3fec37738d841265744ed99b1324977\",\"size\":1177,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-get.md\",\"digest\":\"cad9aee28924775191eafb605ac74439c808d2374cf5eb9dd2c7d22013f75ba9\",\"size\":2181,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-list.md\",\"digest\":\"e4cdf98d4809dc0e4f460ed753cc61c5ab5849793841345031f0b4ce168a20fe\",\"size\":2051,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-update.md\",\"digest\":\"c2d62a867e667e19557d63fa7bfb1222e4546475956f5b8ec6848c16411fdd68\",\"size\":4444,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-create.md\",\"digest\":\"e6b5c0b6492fa6f96bdf0be3839fa2eb3ecde7294253b8409fb9d029f5954b6d\",\"size\":2255,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-delete.md\",\"digest\":\"79973846ea4afc558c349256d51ea1896352bed91bc73dd78fcfb2b85fe4af3c\",\"size\":1191,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-get.md\",\"digest\":\"071e6b2d1be3afbb08dc71bb27bb1bfc717bc1c937a24918b3bf9c39be3ace45\",\"size\":2331,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-list.md\",\"digest\":\"0b83f66ef8615d8d65b6c3d3d362a00ab806bc6375d54b7838b4d8bec7252922\",\"size\":1980,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-update.md\",\"digest\":\"2d4b5f0e63f3e48b7a69ecdcb4d0a96fa1cc5696b1aed0bfdd99cb24f3cc0896\",\"size\":1924,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard.md\",\"digest\":\"2c76cdf7286878885c9b05d80fdf4b2fa9f2a767448f263a9e892dabee052ecd\",\"size\":12588,\"mode\":420},{\"path\":\"references/lark-base/lark-base-data-query.md\",\"digest\":\"c4004baf4533448f696a57a114a0f8dd6c95fa2f7625745fbb46cfdb550ae054\",\"size\":19050,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-create.md\",\"digest\":\"773fd641f49edcfe4c5e1a5b01baaa0e1bee2757176dbf0b8aaa4a4e4969302c\",\"size\":3578,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-delete.md\",\"digest\":\"bdabbaa205ece1016c857e81187825ac056bec8cbd482ab23f02198da024b64b\",\"size\":1407,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-get.md\",\"digest\":\"320cc32a1e4286ecd0595161aff65742eb4e9a449449a57c83e208ff3d215004\",\"size\":989,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-list.md\",\"digest\":\"84f581aa1afe0bd596bc5b5ad6691d2d7b5bfb80ca3d121153d4251b424d2aa2\",\"size\":1190,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-search-options.md\",\"digest\":\"e1b642295d13ced7ea5b28b80f006eee3119cfe34f39e18b11e760cb97749246\",\"size\":1489,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-update.md\",\"digest\":\"3c602ca9b956e640c08f4f48c6b34451e3a20467ee5d35a18b1d4101e8d18651\",\"size\":3463,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field.md\",\"digest\":\"099a943d857ebf8acfaf6d5068b46f8513e73d8f63a0a0b7f566473275bf0225\",\"size\":1554,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-create.md\",\"digest\":\"23ba0d15cab80c788a53d75cf59eb84bd10860c3802bb483020dfac22a0bb8f3\",\"size\":3092,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-delete.md\",\"digest\":\"801058bf1aa463ac0a357562eebb06c4bc3d47f45bc8b6b6653867844f7db486\",\"size\":1792,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-get.md\",\"digest\":\"2c8a267ceca1d400cb0e0707dd5a44c3fdc28ef491c151f901dd87bcf182c98c\",\"size\":2084,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-list.md\",\"digest\":\"7d3df88a25e96ea7d8c15dc31fc2fd4966ec5546e508c9d832255ea48bbc52b1\",\"size\":2480,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions-create.md\",\"digest\":\"7d0a4408bd2152de30aa396976c494d68bd7e75fe3a2d7dc55cf22a3777fbc10\",\"size\":6411,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions-delete.md\",\"digest\":\"c5ea081bdcf5ccb317628b4e43a0bd205a0347e5488cc9d19f9a6968ce5fb6b1\",\"size\":2298,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions-list.md\",\"digest\":\"361932c6c07b1be6f189cb453f15a826f7ef824e925f3abaa77dfe3f063d79ca\",\"size\":2570,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions-update.md\",\"digest\":\"212e2aa37cdce5d49d9444acc6903b3ba6effd3ac51d4fd650dbaa1cf735261a\",\"size\":3708,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions.md\",\"digest\":\"8ad00fe11861df9562b01a9918d9ffeaeee94b79b645bb7da02db411b8a875d0\",\"size\":1254,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-update.md\",\"digest\":\"a1f07ddb86ee24708a6fca8fa7ac3d4ba5df0d9a84943e8775f20c073b3000b4\",\"size\":2807,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form.md\",\"digest\":\"bcc1b20cdc5c962805c61b187e0c317d379a6eb93e89944a1d553486855bcdaf\",\"size\":1187,\"mode\":420},{\"path\":\"references/lark-base/lark-base-history.md\",\"digest\":\"d0b1f46cbd7c837512b1fe6ad931195689bc15cee1a811561646595f1c54fcc0\",\"size\":870,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-batch-create.md\",\"digest\":\"c5879c14c0c589ebdef3ff59b3422dfb7bccc5a0be93d6c663b76f7eda2c11d6\",\"size\":2670,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-batch-update.md\",\"digest\":\"e39e53b0741466c0af0358666078b89cf8da35770df12933b1b698be91b7d987\",\"size\":2943,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-delete.md\",\"digest\":\"35749af2d51d47f95633bb83a08e46d641fe94927e1b726a620caa2537a42e19\",\"size\":1380,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-get.md\",\"digest\":\"b7bb7f7b10efe021736de5240ca50e419e71e44e3710b78a7dc7faecb2d51773\",\"size\":916,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-history-list.md\",\"digest\":\"10741120300d77dce7e382f514e888ed066d4939ddcb583cb641e6b75a4d9d1b\",\"size\":3249,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-list.md\",\"digest\":\"aabfbfec0dd7126397199b428dfdc0d14715561b5ae9b55d9d0488b1d7beb886\",\"size\":4286,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-search.md\",\"digest\":\"8d18000e59c11774dd95017356a4da1c72546cce4353ba8df0cec4667d3b2468\",\"size\":4263,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-share-link-create.md\",\"digest\":\"592c324f47df6a61084a8fe9ba327724a3e7aef37eb855d6cce0b216060eba32\",\"size\":2208,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-upload-attachment.md\",\"digest\":\"6f3e29aa9b8281405faff37a80f22c66dba72939c6061e973dbc805bb601f3f6\",\"size\":1959,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-upsert.md\",\"digest\":\"41846e79fbb70ce966e7ce7c45c6d3efa3c1e022354c8bf3315ed08b4fccd8be\",\"size\":2666,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record.md\",\"digest\":\"060d5074993435ceb5cb21b8f028fb5f7b4274c47262fe8493d5fcf955707d0b\",\"size\":3705,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-create.md\",\"digest\":\"966a82d2e9608686849626b72ff099e0facfec1f647f7aa64174c8dc3cda570a\",\"size\":3650,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-delete.md\",\"digest\":\"3b23ee86d43dceb281282411df36e15c5925d2442b4378a6f633420f0f060716\",\"size\":2387,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-get.md\",\"digest\":\"8851569583e1d8d61e8ebbd92d4f48091e1a6a0336fcd970af63882941448054\",\"size\":3307,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-list.md\",\"digest\":\"ca5bc190347b02307f18ee14424b50aee62cd6cc574886cfb8825927bcf029e3\",\"size\":2933,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-update.md\",\"digest\":\"aceae4ab238a47b1d73270f65bf5023a3e89cdf70236263266e03a6eaba56972\",\"size\":3166,\"mode\":420},{\"path\":\"references/lark-base/lark-base-shortcut-field-properties.md\",\"digest\":\"85abdc3999e38376d2ef5e3e8f8530fc57fa6b6486144b9d438a8823e05369e7\",\"size\":32026,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-create.md\",\"digest\":\"159704269dc92d584f57febef511a9a8eada96b6fe85eb94193383b451a442ad\",\"size\":1843,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-delete.md\",\"digest\":\"c101e61750e9312929cf8f830ae2b494df6e12965dd47685c4f9626822756679\",\"size\":1414,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-get.md\",\"digest\":\"74123d055adf7725244e48f5646b8ae26ffdcc04a87e743f3098ed0b0a8ad6a5\",\"size\":1480,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-list.md\",\"digest\":\"670624a861c3e66d0525713cb72d5fbe26a789bd883052df0388bfd2ef961d99\",\"size\":1103,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-update.md\",\"digest\":\"7b4115971b9aaa88be8cfeb068e4ad0df91f4fa60de8ab9355810fa89b636c97\",\"size\":1088,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table.md\",\"digest\":\"e000075679e5d8b9d51e2908c26d34d4e9aa4c89c14f788253196debdc19365d\",\"size\":1277,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-create.md\",\"digest\":\"81dddf4bffd7d805c277cbb8e2d8d3f51247bd20ae7c987bf4a1b8b2515c5a29\",\"size\":2582,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-delete.md\",\"digest\":\"6544b0beb80098cf16487e789050815b1af4a7d6e8ece9db8da37e322788ec67\",\"size\":1061,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-card.md\",\"digest\":\"215970e06056af1b73cd34886fa8ab4ace8991e14308bed142a010b8f8999a78\",\"size\":877,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-filter.md\",\"digest\":\"bcbd9e8df79ce9d24e562d09fc3ec513da6738d890fdcd525c45a0f318290038\",\"size\":883,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-group.md\",\"digest\":\"24a005d1ad8bc61efd7b8de5627546388787b7b31a23509f7be6d7f6a7b4e51c\",\"size\":880,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-sort.md\",\"digest\":\"1bfb13dacf7ecb9305535ccf198bfe844c583199544c7458499b5ecefc729be4\",\"size\":877,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-timebar.md\",\"digest\":\"aec83ffd1c7126304553e358bb4f06d62dc2fe3c422336102067c7e535f138d7\",\"size\":892,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-visible-fields.md\",\"digest\":\"e29d32230c791c2f21bec96b8e145d0426219fa5669431848ec0259dc68c8cbe\",\"size\":922,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get.md\",\"digest\":\"1a7866e35f66e9a29447cfcbbb42b8127f311471c5106a5ab32af4dd5e20b33d\",\"size\":880,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-list.md\",\"digest\":\"791157d8afa07361b03f6cc79e4737d31e2bb763824d363b726abc9e42c0aac0\",\"size\":1183,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-rename.md\",\"digest\":\"a048be3869d620118c27347b6b1b0d35300fa5a08a4625509cd5b906f3ae1d39\",\"size\":1038,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-card.md\",\"digest\":\"0ffb20fcdae884503c9940ccbc3f6556efd70804d11e12da1924df9fe7ce4dca\",\"size\":2251,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-filter.md\",\"digest\":\"4dd7804ce91772294169072c20d3fd2511baad6664ef385ec4370402de13b213\",\"size\":6534,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-group.md\",\"digest\":\"42366cbf8cbf1e2e3a048b6e569880734ca14ba63bb69c052cbcf53e53acde14\",\"size\":2403,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-sort.md\",\"digest\":\"ab7e1ad8070c28b1fdeaf827eee59d73584f4973705912d61345ed3e052beba9\",\"size\":2291,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-timebar.md\",\"digest\":\"704528a273a50e4d93013ecb5a6e02f47cd2e1c1b7ca5f8e66d05748652b3ee9\",\"size\":2736,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-visible-fields.md\",\"digest\":\"c50d32e6dce25f4a71fbb67eda7dcbfde72bd3d90b3c80ebe26321b0f1fe50d0\",\"size\":2180,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view.md\",\"digest\":\"32402cd5083e8b37575a855d9e9fe7e6fb5dc6da9b13e8479d49f211b3115155\",\"size\":3453,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-create.md\",\"digest\":\"ab54d32128bfd6a89319485d2f4c25663606b9a6d31740965e9f693ab9f6b534\",\"size\":8878,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-disable.md\",\"digest\":\"911ea146d1fee02ab1d4fd53080874dc0b94b38a969b9356e9598466362f398f\",\"size\":3156,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-enable.md\",\"digest\":\"d499ae0b703875fec0406d8f11775399a3c3b13d0bc9003c72c76d469dc73641\",\"size\":3148,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-get.md\",\"digest\":\"b926f557d9aa019bfeb30b186ca631c8a58b4042a11d5370788bd6059226cc14\",\"size\":7464,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-guide.md\",\"digest\":\"5fa91e39b21e3c3f2a32b96c0eb0edcddc64f9fe903b318615472986959fdb47\",\"size\":26427,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-list.md\",\"digest\":\"f0433cf7dc40dd8eb3bbbfee994e760ffd017e302031818120383d35edfcadef\",\"size\":5212,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-schema.md\",\"digest\":\"89d8610ad1ee3312efae8524112e9c7ac6613acc2c2a802c34376627ee823081\",\"size\":49139,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-update.md\",\"digest\":\"7bbb8baebc4bfcd81f34df5935bb8b757e94005d2f1c5597379dd2b58a4d5cd3\",\"size\":7553,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow.md\",\"digest\":\"11bf335fc68bef918a79664e9185a1694c29556315fc9531ba308fdf0ae3a9ee\",\"size\":1997,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workspace.md\",\"digest\":\"812be3c7ace8bfed55b615e4bbd8516b464c76fb6405479a454446bde6bbf0f2\",\"size\":845,\"mode\":420},{\"path\":\"references/lark-base/lookup-field-guide.md\",\"digest\":\"6382f4ed07b01a851243814854b29a98b886a052c68a840407ddcf97d4a59e3d\",\"size\":21979,\"mode\":420},{\"path\":\"references/lark-base/role-config.md\",\"digest\":\"19316f0a26a87706a7614df1b565880f00d857c561c666ab8a0dd53943ef7cd3\",\"size\":25503,\"mode\":420},{\"path\":\"references/lark-calendar.md\",\"digest\":\"ad558e85f381c2a20f8f8d602b17264af0e1f441348bbbacdcf3afe7d31c7561\",\"size\":15421,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-agenda.md\",\"digest\":\"c2054ce3c43ecbe88ea44403e9e164f7f1c7d9b2ddb1800abbd14f3a09bdac04\",\"size\":2848,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-create.md\",\"digest\":\"c6240fafc05bfc073f6ceda6d978acc98142bbb483f4f5f7c6a682e756f6691a\",\"size\":6333,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-freebusy.md\",\"digest\":\"810a6c2c005ca45e9122fca5704c5cabf251e50f88c6b6afb6361d677a0d1b0c\",\"size\":4617,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-room-find.md\",\"digest\":\"fdff4b26e8b8e4dbe33272389758cdc497552731ce783bd2b3f1d2d4fa285fad\",\"size\":12490,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-rsvp.md\",\"digest\":\"0c7e09fb2f1d07bc1bb06e033689cc64b615c38e88ab5209ba301913a12c96e4\",\"size\":1726,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-schedule-meeting.md\",\"digest\":\"4773c4fb81859d0643c3f7802d9c02a2b57bedc16f9902ad551a4d8e782be117\",\"size\":19081,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-suggestion.md\",\"digest\":\"9cf2d32f4ee60b2fb814de1135972140e94a9bca618f593183c9b47346b5f25d\",\"size\":7910,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-update.md\",\"digest\":\"967c398d9d4816a5a555e5af01d27bb514c29aa8e77548b7fb02f72ed97944f0\",\"size\":8433,\"mode\":420},{\"path\":\"references/lark-contact.md\",\"digest\":\"7386d4a80c881447231dabe03de74a073fc06c38a8f37bf23ece1e8fcf1b255a\",\"size\":1322,\"mode\":420},{\"path\":\"references/lark-contact/lark-contact-get-user.md\",\"digest\":\"47c7190638d40fd5071b65689e2b99609d61b3d9ac1a302225401ed17deb38d6\",\"size\":1771,\"mode\":420},{\"path\":\"references/lark-contact/lark-contact-search-user.md\",\"digest\":\"072d092c96f87b48a3c4974202eae361d7c1ac10f6edb0b437806c987e0c0dfa\",\"size\":1499,\"mode\":420},{\"path\":\"references/lark-doc.md\",\"digest\":\"a43c74ed2b0248dcde9d0a16a780e1c7a90c2b80a5b872bfdb03940c575d787b\",\"size\":9199,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-create.md\",\"digest\":\"26360050ff601e0e6a0b2e24e2fd678bb955f0617606b7149d83e764f72f61c7\",\"size\":5464,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-fetch.md\",\"digest\":\"e23367bcd0795f5e14bb8a1c2d647bebc26ae39a84c762eff79411bd170b1226\",\"size\":9673,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-md.md\",\"digest\":\"e4108795236bdc0c8d1fabcb8652b521aed0f6fde5e5058473e969bc880a5837\",\"size\":5987,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-media-download.md\",\"digest\":\"58bc49188b05771fd247ae744a5529bd67d551c3baf45c4ae85e7df907629017\",\"size\":2350,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-media-insert.md\",\"digest\":\"649934b2a1458b51888b5c7fecfb7c8cd32069ecbeca51f0195b5ff9475ad91d\",\"size\":7525,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-media-preview.md\",\"digest\":\"27ac73168ef601c7ffd4eb4cd2a4d7c8f6ee1058d1498ae75e4613d175a3194f\",\"size\":1743,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-search.md\",\"digest\":\"7714586777209b57b55bcc1277bcc828fd5b5a11e456f28b444f109ee6029ef1\",\"size\":17975,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-update.md\",\"digest\":\"4b3357cca4940150deb5acc359fbf9e3d3e98d679537df450b29f4c716e30455\",\"size\":14595,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-whiteboard.md\",\"digest\":\"d9dd5922837a756ac9448e545c13253aa3d269e3a0215ec68bc683ffbad411ac\",\"size\":3717,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-xml.md\",\"digest\":\"01709b80873ea512f3aaa07bfde6aa7f2b360cc9fa6a2b6e07255151108286f3\",\"size\":9143,\"mode\":420},{\"path\":\"references/lark-drive.md\",\"digest\":\"4ac2f567cb351cd36c0ef93ca27b3a0bfa92fb162aa546419da1bfaab255e2f0\",\"size\":23520,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-add-comment.md\",\"digest\":\"4443d6b6918aad75a878ed8a0c30901d30d9f7459002c94aecc28abea5a90e21\",\"size\":9748,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-apply-permission.md\",\"digest\":\"f6ca8a77155bd45d9bb8d3ba5d66b9af0e4df0b7b33e16f03e627adc8cc53897\",\"size\":5461,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-create-folder.md\",\"digest\":\"584d64754b35926b993ae66585ed8f744f4b14466bb4270eaed5a450e15311d4\",\"size\":3745,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-create-shortcut.md\",\"digest\":\"9ff53f2e77fd96aa3eeed52934ad9997332047430cf803c464abb65c6e74f3f5\",\"size\":6090,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-delete.md\",\"digest\":\"2d6aa2b941eeb472a77bac7c3798e77ff285b507a773234e4228278d18785155\",\"size\":3047,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-download.md\",\"digest\":\"db665ac0a1210d48dd9c7d4bad018a36ba4860aae013dd01600be45d9f627bed\",\"size\":754,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-export-download.md\",\"digest\":\"5d0e13f82dd2a8ad91caa02f4222cb8469931479da65a99fab9c8270b2df8249\",\"size\":1616,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-export.md\",\"digest\":\"a4b0f8487b12c5891e616dc6e4371581950339a493e342f026922e37771413d2\",\"size\":3593,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-import.md\",\"digest\":\"3442fbe42141d7a34d85b2524858b3eea4d8bd84805d4ab7255c62385669d535\",\"size\":9876,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-move.md\",\"digest\":\"719c3c9f9889f0a69bcfbe1dd7f6fa013764c90f23fb0f03b0b3ff78ab9fdcd2\",\"size\":5547,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-reactions.md\",\"digest\":\"7c6761ac97e2e65f1ddb9a368a916a2a2b17b2d7c8b3146f1cad531b394c720d\",\"size\":6271,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-search.md\",\"digest\":\"a7449faa1e7bbff03082c0851a23e7e7ab3496b51e3d86076274e98b739df591\",\"size\":19496,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-task-result.md\",\"digest\":\"b17c30a6481924fdbbad8bee4a50da678b4ec9c708aa886d8e4b725af73c9ddd\",\"size\":12191,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-upload.md\",\"digest\":\"3fc4f91b080111e0198b59d2b76bb17529231a18cdb929215a068283958da0db\",\"size\":4784,\"mode\":420},{\"path\":\"references/lark-event.md\",\"digest\":\"5fcd0040bf6a95dcdc96072ece510703d5d7ae4ef5e78c88b7f297f68a2dbb31\",\"size\":9685,\"mode\":420},{\"path\":\"references/lark-event/lark-event-im.md\",\"digest\":\"4ef2de1f09e03f70c2549dc8de4ba3d72e6df6b4262b94eaca52666107e3835a\",\"size\":5480,\"mode\":420},{\"path\":\"references/lark-im.md\",\"digest\":\"124f0668b2f72c8c093997c325ad725ea62d408e54390950d6eaf390fcf080cb\",\"size\":12516,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-create.md\",\"digest\":\"b8bc93ed6050f57780c5cfdba89a22b03657f24dec398f8da6dc94e8235c21b2\",\"size\":11193,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-identity.md\",\"digest\":\"80d1d1ff90ed7c2583c8bd5d203eed0fd411d68f44f7d426cf2d07f727e84bfb\",\"size\":3596,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-messages-list.md\",\"digest\":\"09cdc14e510e5b76995ab8a09dd0d9958395e267246ffd2c9bfc88c08cab421c\",\"size\":12016,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-search.md\",\"digest\":\"243571538c313d09d7f740ab6410f1126584be7953b6150a8026f7d48cacc9b3\",\"size\":9770,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-update.md\",\"digest\":\"607bf1293c5c0d56db22647a1b1f001461c635c7ebf1cc607ae550705f225aeb\",\"size\":4041,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-mget.md\",\"digest\":\"dab7948007922274ad6a0b6532571e5e52b663cc9c5d8943d31788a24f0a07dc\",\"size\":4445,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-reply.md\",\"digest\":\"a01d4430af97ef6099b674a9deaffa405d79ec872e96746ef26acee0c71667f2\",\"size\":13773,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-resources-download.md\",\"digest\":\"f75d8cdb67cd5772b27da7bcf9482da5be3ce12fac2daf34de4c7bdcd37fe357\",\"size\":7625,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-search.md\",\"digest\":\"a90a02bcc8cbb1aa0dcebf4baa6f54ec15c2dc478505a3ce3dc4fe3bf5a6dd63\",\"size\":14743,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-send.md\",\"digest\":\"ac421c43664f90fe027d8c5fc210ab1e2ad8bdc219c42cb6edae14627f267aa2\",\"size\":15788,\"mode\":420},{\"path\":\"references/lark-im/lark-im-reactions.md\",\"digest\":\"b9af652a41a73af6a84acd83dddbdbab700a3b44655bbbe2024388da223fb106\",\"size\":12015,\"mode\":420},{\"path\":\"references/lark-im/lark-im-threads-messages-list.md\",\"digest\":\"89b9912f3ae856c6935ef51e954ae8444dbf2b4a9c62a8fcf02361bb2afa3651\",\"size\":5841,\"mode\":420},{\"path\":\"references/lark-mail.md\",\"digest\":\"11cb64efe6d11ebe8b9751d3e57a285f3e21d255f8a507b0695deb2e10540f0e\",\"size\":40944,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-decline-receipt.md\",\"digest\":\"267fe11794a9d41eaa848607f7268dd8cd550dbacd5789fd9f4538e2b7d7548c\",\"size\":4180,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-draft-create.md\",\"digest\":\"ffcafe78fd925afb97e2da61f8f8c5ac189bd944f5be07a04e28a2d2d6640a84\",\"size\":10075,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-draft-edit.md\",\"digest\":\"8bc23f56ccdefa2d6e0afd10978c157639ae14646bb2d5b1ebd0bdc2ffe6afca\",\"size\":19802,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-forward.md\",\"digest\":\"36033f1af3bf5b896463a740ef087cce53ccc02e5401a8eb51e17dd84fbd0a81\",\"size\":14531,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-message.md\",\"digest\":\"3cbfcf78446212acc1d49fcf5e02fa904797daca275c5c8fdd3410766f510e1e\",\"size\":13467,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-messages.md\",\"digest\":\"995fbb637dcdca97991a0520dd8310dbb84cd6ac22903679ee6d674991c7cd9b\",\"size\":5139,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-reply-all.md\",\"digest\":\"2d4032859fd58b6a3cd584211de89f61d0f57c3bdec7bf2394cd12a355ee622a\",\"size\":13885,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-reply.md\",\"digest\":\"f5baa3e0f44111237437533c96927143df5222452f0c9d2ea0f68a31ef2a9889\",\"size\":15390,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-send-receipt.md\",\"digest\":\"cdbb185cf4973d80f6186811d0b76b4cfcc4285fd6eb047c407ef9a603a08ad1\",\"size\":6424,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-send.md\",\"digest\":\"45a8862842fdf4a803a00845ac1c60499aadb22590815f732bb7b8b163407272\",\"size\":14270,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-share-to-chat.md\",\"digest\":\"ea94601c4c747ef1cb9dca8182e73a3a82377a28124d6e75f4125025fbc5e8c3\",\"size\":4537,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-signature.md\",\"digest\":\"38c0378777fe466cf9406108d2f00db3015b623e55bfc23c15cb3d344cca4928\",\"size\":3199,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-thread.md\",\"digest\":\"dff999f72264c64e43d93f55e67cfc6da9089fd3546ccc8cfc83889bed4749db\",\"size\":4868,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-triage.md\",\"digest\":\"858ef54e2d80936aa72b741af94d4b1cb6c74f3b2f8b3a0ae540ef834601b121\",\"size\":8290,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-watch.md\",\"digest\":\"b30ac05ff2f9ce8b807006e9744c8e4e38121c1ab0d38004faa3f1f8e9a96937\",\"size\":5681,\"mode\":420},{\"path\":\"references/lark-minutes.md\",\"digest\":\"d92e5f73a2f15380a4611e5623f9163e00112833dfaeba69db2fa7e9e63f8eaa\",\"size\":7401,\"mode\":420},{\"path\":\"references/lark-minutes/lark-minutes-download.md\",\"digest\":\"80a2aba753bb5af47673c8b8b71290e1e57f6e79e3623603241a0962f3ddbb54\",\"size\":7470,\"mode\":420},{\"path\":\"references/lark-minutes/lark-minutes-search.md\",\"digest\":\"1f0fd7568ba574b6a7b2a499cfd25a7c42925e6808c71b8705507500525a4aaa\",\"size\":11683,\"mode\":420},{\"path\":\"references/lark-okr.md\",\"digest\":\"d33520bcc20579556a19ca5bcb2cc1fbeb80638af38f1560198d92d4148b4b48\",\"size\":6292,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-contentblock.md\",\"digest\":\"f65a3b210ec1b3ac7c2aba359a3b6692f4c1c87335379e76772f920284105fe8\",\"size\":9604,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-cycle-detail.md\",\"digest\":\"86a3dc4018748426def1ca118a9eed6321de77ea4be458951007b7f9fa6352ab\",\"size\":2560,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-cycle-list.md\",\"digest\":\"de492e806d6bfa23b562d90f39aac9a43cf3b93d2651f227e9f3680b6d58134e\",\"size\":3882,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-entities.md\",\"digest\":\"a51d7f4f85b0ed6ba36e95b04af45aab7f74fa0cec87e84971beed8b06086a09\",\"size\":21859,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-image-upload.md\",\"digest\":\"777aea7c0d562fc2014f9dca28b4f910f1a58f0e9dee487a19eb324ad31276f8\",\"size\":3554,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-create.md\",\"digest\":\"811a78881300565c41287bd3847addfb9c67064024b26d1269f3455a759e689b\",\"size\":5836,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-delete.md\",\"digest\":\"a0a4b9f124a78f29b8e505a842abc01d8194bd6dceb8add0a5cbdd8cece8f4bb\",\"size\":1489,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-get.md\",\"digest\":\"f849c4a8f3b3abb213dde6da83019ee218ec5d40e16c9cf7dc00c6421efabd02\",\"size\":2186,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-list.md\",\"digest\":\"ee1a2ea7df138fb662b8769c9a2792c95561d43b71ca57f275777a2b2a6ccbee\",\"size\":3282,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-update.md\",\"digest\":\"72d8f99dbf95c615d72cef192fec9522b4933881290e336504d1d7282d4dbf8f\",\"size\":4966,\"mode\":420},{\"path\":\"references/lark-openapi-explorer.md\",\"digest\":\"822cf3dd71610824a640e8bf67354c22b94cfc78ec701c648287cb2dc7290b2c\",\"size\":4982,\"mode\":420},{\"path\":\"references/lark-shared.md\",\"digest\":\"13657eb5c89ac6e5c716cfd9ea1e8c0570abb2edc332fbc53b26ff38a7bf96b2\",\"size\":4242,\"mode\":420},{\"path\":\"references/lark-sheets.md\",\"digest\":\"e3a8854db2d016049de9bb71916197a992fde724a96480c1a8b5e75816868054\",\"size\":20152,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-add-dimension.md\",\"digest\":\"0c46243beb929a7358d52b080bff434925385daeb061d5a20950f7d84bad0d2b\",\"size\":2077,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-append.md\",\"digest\":\"f92479bea20b8ab168f3d0474d7dee3c64b95dc75555047eb809bf3acfebfd80\",\"size\":2497,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-batch-set-style.md\",\"digest\":\"4645b41ea31184fc5dc59de20a34a5924813c34dc65a5225750845fdeab0fdc2\",\"size\":2324,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-create-filter-view-condition.md\",\"digest\":\"0506659b45afe666bd0d3415e653489377a5b0094c9df8c2a1c21f7f2e718eb9\",\"size\":1970,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-create-filter-view.md\",\"digest\":\"0e408421cab5eecf2aa04e39e479b8d6f6bc037105e6e1f7699f8872342120c9\",\"size\":1826,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-create-float-image.md\",\"digest\":\"0419d50105b5816dfbdf737a878850bbf5c4d5cc1e10611ca9d614ebcff6f51b\",\"size\":5321,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-create.md\",\"digest\":\"82dc3587acc5adcc0ace8798ed81c1ed25e5e49683d0142b407e7db4158ba9fa\",\"size\":3021,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-dimension.md\",\"digest\":\"e6b59217bb3969d9f4970ca0fcafaa6f94072d2b78ebd3e489f4ee7eb7aafe18\",\"size\":2369,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-dropdown.md\",\"digest\":\"0ba323acbf046ba64695942ed55241ee398c58cc120f6d53cd22ca485d174fa8\",\"size\":2011,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-filter-view-condition.md\",\"digest\":\"3b1a3431125d7eb9459e13bec68e9acc7c5e14627dbaff02105dd01b85550e68\",\"size\":940,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-filter-view.md\",\"digest\":\"7bddc0c1b49c9ddf9e6718a85432f04239642cf31a6008de06e02f67318990b7\",\"size\":844,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-float-image.md\",\"digest\":\"e6e66e583552ab0b7e926b81b5f0b6f8c939c26ea01659bb2d6ae6a994cda193\",\"size\":1430,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-export.md\",\"digest\":\"466f59db8922f0dc7e89e84a36f2bf7527fd26d5a0d7063f832f9fc372cab002\",\"size\":2269,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-find.md\",\"digest\":\"d4e1864c082f78ee2e8808346bb19b81915cfd2c40f0110dba8d0e510a5b334a\",\"size\":3521,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-formula.md\",\"digest\":\"576bff8190cb9c86da25432a7fbed3fc59f238c2433ad4344d69e1fef145cba0\",\"size\":4233,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-get-dropdown.md\",\"digest\":\"da68446a6b43bde30fd444730cbc6ea7f0763a962130590d28e0df9bb20df57b\",\"size\":1740,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-get-filter-view-condition.md\",\"digest\":\"48987068ede198367c92dd6c2f62da7c1d8c32bdff38b48d16bd181ba001c1a4\",\"size\":947,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-get-filter-view.md\",\"digest\":\"e1cb58bd4bd03f2d7e76933177af5ba7e5a0080b6363903cc80fd1b49c40c2dd\",\"size\":807,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-get-float-image.md\",\"digest\":\"003f0450fc6773e2958d9d170c65124d6dbe0d2f8839bd9c9d1445dacbfae557\",\"size\":1732,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-info.md\",\"digest\":\"84867505a9f3802e313e5c32b05f3245f0b1e4a739fb7c27287a3c8b018afa97\",\"size\":1600,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-insert-dimension.md\",\"digest\":\"ae927c95deeab3d72e98492d13528c9ae554d10800bde2110cb50044bf0b3aab\",\"size\":2680,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-list-filter-view-conditions.md\",\"digest\":\"9674249bd0f26877bdfbed8fd647d25bdb0eb9d931843718d42f6faf922c8388\",\"size\":893,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-list-filter-views.md\",\"digest\":\"e04004bad2b4169e6fdf0b3505194b28dc736cf2232067613d9197a8f0a2b084\",\"size\":806,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-list-float-images.md\",\"digest\":\"a3c48c8c880528e6121cfd4229e5886007a7e51b3a9c109010122d2e2de46040\",\"size\":1602,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-media-upload.md\",\"digest\":\"26605e1e76220021f35924f5ea0988563d6d9ecba6fda65de9feeb2890e9db50\",\"size\":3575,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-merge-cells.md\",\"digest\":\"a46a7c94fe94f43eb830b63409bd9542e42f3db8605e02a898d6145c0c8941c1\",\"size\":2199,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-move-dimension.md\",\"digest\":\"cd42d7b4ae679bf968d2fc37956154e45cbd4a4ae03082453156c70352d1857d\",\"size\":2341,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-read.md\",\"digest\":\"8b5db888a721ae96f31e7da2b0537dc21796260b39072ac8c422e5f6b4fbb441\",\"size\":2917,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-replace.md\",\"digest\":\"cc34e86d9ba12e7575db3fd0c5006f9318cb15070e98cfab8a332dff1eeb763e\",\"size\":2854,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-set-dropdown.md\",\"digest\":\"6b8246c109e3c537ec17f3161f633cef512598422e4973e18ba3f5342975a802\",\"size\":3264,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-set-style.md\",\"digest\":\"f095783161e36676f85b46b0c31308b66a80db6dfefe65c4c8c54242b2bd6898\",\"size\":3764,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-unmerge-cells.md\",\"digest\":\"5ca41afbd297cb4db20a0b50c0908c3e890b0b57801cc33f82ea988d2536e106\",\"size\":1857,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-dimension.md\",\"digest\":\"83cffdaf374904e588ded164aaa1de698fc18b18615d416631e9781173a829a9\",\"size\":2838,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-dropdown.md\",\"digest\":\"bb5c8e6b444065f4a3555718c3052dd2728ca9061fa069fe9545661a0a347a2e\",\"size\":2444,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-filter-view-condition.md\",\"digest\":\"b61b8729a9601168f58effef09f2027c5a564575019fb748da42b31d6939cc16\",\"size\":1109,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-filter-view.md\",\"digest\":\"3e735566fb0fa1b3137678d72ff37d1d7904990544d2ccdd59c82079ab340204\",\"size\":865,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-float-image.md\",\"digest\":\"f9a3dee2aa3d9091685d5f6059758f7f78b45088dd1e05b1d4b8e99a8c332341\",\"size\":3689,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-write-image.md\",\"digest\":\"e2b97d27369e50b4d7d609eb22232672a52de37e1baf8baaf5318f396ea9ae12\",\"size\":3000,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-write.md\",\"digest\":\"d4e6e2136c3559100b4ee010abad78a5101db62010f0ba6f656ad310dfbeccfa\",\"size\":2802,\"mode\":420},{\"path\":\"references/lark-skill-maker.md\",\"digest\":\"13cb0c86eeb0747be0cdd5de05bb15cd2a29c81cc56c3aa9a9211dd4de82d3eb\",\"size\":2451,\"mode\":420},{\"path\":\"references/lark-slides.md\",\"digest\":\"ac5710259afe91d7fd0a538a605d45dd01362d66bd1ab5ab7f99dfa8d6a742ee\",\"size\":31154,\"mode\":420},{\"path\":\"references/lark-slides/examples.md\",\"digest\":\"3e465212249d8f7ec45460e009d3008aa3995824a82cc0840d81f37c9b333ff3\",\"size\":8657,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-create.md\",\"digest\":\"310fb9c269875b43ea5e4bfa29b9822656fb1a68968a2e43f182770889379508\",\"size\":7797,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-edit-workflows.md\",\"digest\":\"657ee5e2560f92a917d5ded2bf27709d62c37ed7e5e3473210d51ec520da9143\",\"size\":9054,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-media-upload.md\",\"digest\":\"156ef086de0c7cc3f41b94151bfe8a0c6b3871a0ce91b71646641abdcd4fda2c\",\"size\":6899,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-replace-slide.md\",\"digest\":\"5882afe329e99df2286950ee813a94c10dece90b07a99288686703587cb54094\",\"size\":15837,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentation-slide-create.md\",\"digest\":\"67af18aa008b7b31676d336000a6e1d1974a3158cbfad59d20f22293e4a6d1e4\",\"size\":9145,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentation-slide-delete.md\",\"digest\":\"aa6b82f30d476cb171e0ba7f98c60988f10936cbb602c110d8a902baa22abe1e\",\"size\":4594,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentation-slide-get.md\",\"digest\":\"3ff08d9def7e6ef69f5fd5a8a417b14700203b289b71821ac575e4cd07c026cc\",\"size\":3985,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentation-slide-replace.md\",\"digest\":\"9de02710406b7eb3c898fe3b715fb8334b3b9860217f19a0e78174f04cebb743\",\"size\":8735,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentations-get.md\",\"digest\":\"67e99dd7601ba7de292448af1f29e2bdbb26829cb969797e57568660ece0a5b9\",\"size\":3425,\"mode\":420},{\"path\":\"references/lark-slides/slide-templates.md\",\"digest\":\"73c50b89cf5c7b2ca5408161df75d356fb62059d720220b9352d676e7a875e1a\",\"size\":10182,\"mode\":420},{\"path\":\"references/lark-slides/slides_demo.xml\",\"digest\":\"161034420a40441ee49cddeb04bf4c9381617651827be97fd959d054afc94d7e\",\"size\":12898,\"mode\":420},{\"path\":\"references/lark-slides/slides_xml_schema_definition.xml\",\"digest\":\"b2d8eaa71acb58786957a951c8110488a25d10b82ec32caf7e23aaa341cc0a32\",\"size\":173099,\"mode\":420},{\"path\":\"references/lark-slides/supported-fonts.md\",\"digest\":\"0ff1f9503744762c00d427310db1eed04ad556a5f0e18d7d1c2fd5c83c5bc9d6\",\"size\":2143,\"mode\":420},{\"path\":\"references/lark-slides/xml-format-guide.md\",\"digest\":\"043cebac11862bc4e6181f36876bce00faf88b36b7b1c3e220485e5844779b03\",\"size\":10990,\"mode\":420},{\"path\":\"references/lark-slides/xml-schema-quick-ref.md\",\"digest\":\"6fe9b8baa6270ca0abe2d57cf3cf2958be3e69bc0c6518c32616322772c821dd\",\"size\":7434,\"mode\":420},{\"path\":\"references/lark-task.md\",\"digest\":\"06b6799eac93a974a1eb3fb7c35d89a5a90e93c158c300008d19d4dec5fc8fb8\",\"size\":9925,\"mode\":420},{\"path\":\"references/lark-task/lark-task-assign.md\",\"digest\":\"5345d4b8784f7fee9270865ae52d4f83187032c8971823dd6970667f1edcc3c1\",\"size\":1684,\"mode\":420},{\"path\":\"references/lark-task/lark-task-comment.md\",\"digest\":\"3ba20bd078272aee771e850219324ab5c0bdf6e77e10e592e36b6dc2b77f358c\",\"size\":1317,\"mode\":420},{\"path\":\"references/lark-task/lark-task-complete.md\",\"digest\":\"cef972c53f76c4a4b312732f7ea5532692d9e194367339b5f6a264c2441d9969\",\"size\":1031,\"mode\":420},{\"path\":\"references/lark-task/lark-task-create.md\",\"digest\":\"b7aae12593c7972ecf4e6cafb5e7117cd9a81d5e77c8ffa3ba94e2827f639a54\",\"size\":3564,\"mode\":420},{\"path\":\"references/lark-task/lark-task-followers.md\",\"digest\":\"820bfa6ebd3fe1192eafa901fc3b29d14cef1f3ffb63a2ab56d691fe4fb1d51d\",\"size\":1549,\"mode\":420},{\"path\":\"references/lark-task/lark-task-get-my-tasks.md\",\"digest\":\"784c71b630cadbfdabf0103263109f9e326de7c124b7379e3accaee7e3d2d8fb\",\"size\":3863,\"mode\":420},{\"path\":\"references/lark-task/lark-task-get-related-tasks.md\",\"digest\":\"0fb9365dfc5875b0ad2b63bb9b7d07f7d7c695798b8587c257d97170c90219bb\",\"size\":3521,\"mode\":420},{\"path\":\"references/lark-task/lark-task-reminder.md\",\"digest\":\"ec9b79fa0324fcdc82e30a5b1aaf9d273fe24d8a5469f4724da16e1f28eaa575\",\"size\":1921,\"mode\":420},{\"path\":\"references/lark-task/lark-task-reopen.md\",\"digest\":\"2b350a0fab8e4bb3664264082529b31f45e2e3b2b3e190b6dc7ab522081fe8bb\",\"size\":1027,\"mode\":420},{\"path\":\"references/lark-task/lark-task-search.md\",\"digest\":\"912235f157b012e527b1d1f4b2a00fb0129cb5abe6e7c3543e17433ee84a63af\",\"size\":2196,\"mode\":420},{\"path\":\"references/lark-task/lark-task-set-ancestor.md\",\"digest\":\"c3a947d8394eb04891a636bcd3edb65550c08fb6ea38750af7436b80ae003471\",\"size\":1116,\"mode\":420},{\"path\":\"references/lark-task/lark-task-subscribe-event.md\",\"digest\":\"7ade9e0d5d225c93f53332a8b07ac8c160587227c7891e8fdbdde58d9b2fedc7\",\"size\":2816,\"mode\":420},{\"path\":\"references/lark-task/lark-task-tasklist-create.md\",\"digest\":\"548343a71b17d3da7ea8c53078fee6904819ef7c1450ca878859aea90523e9fb\",\"size\":1578,\"mode\":420},{\"path\":\"references/lark-task/lark-task-tasklist-members.md\",\"digest\":\"f8b6bc395ad67e46b02d72b1595845e78bfedd9d1aa13ef7723b1a35350ad0b1\",\"size\":1524,\"mode\":420},{\"path\":\"references/lark-task/lark-task-tasklist-search.md\",\"digest\":\"cf10c6bffb80391c01313e45b1c2121886e99d0bf88b32d126379cdc579fe83a\",\"size\":1891,\"mode\":420},{\"path\":\"references/lark-task/lark-task-tasklist-task-add.md\",\"digest\":\"a969be95a6f6ae2c32f575bff5b2ef6734fd18ae21fb4fd7ae8909a1b23ca410\",\"size\":2126,\"mode\":420},{\"path\":\"references/lark-task/lark-task-update.md\",\"digest\":\"751d88b63248b363cbc722d977e4ec5a2ac8a9b0546f7424d7f9398516ca7a75\",\"size\":2534,\"mode\":420},{\"path\":\"references/lark-vc.md\",\"digest\":\"bcff228438452ed5ab1432f5ee1a3ae90e76f018dcc635cecc6a24879d739cd1\",\"size\":10063,\"mode\":420},{\"path\":\"references/lark-vc/lark-vc-notes.md\",\"digest\":\"0c56efca3a2b240ee5a2d629f9341ec67d87af600b68d6be4bdaf0a2b7bef6ec\",\"size\":9945,\"mode\":420},{\"path\":\"references/lark-vc/lark-vc-recording.md\",\"digest\":\"6931c153247eda34ba726ff4657579bae1c2051ee8d003e89881a66eda1c1147\",\"size\":6728,\"mode\":420},{\"path\":\"references/lark-vc/lark-vc-search.md\",\"digest\":\"f8c81f0ed85a1be78035be719e6de022aa44d0340069672e3ed38ba0b8eabb93\",\"size\":7745,\"mode\":420},{\"path\":\"references/lark-whiteboard.md\",\"digest\":\"dd165ae69ce1d82c74be8579ef658353ddce11bf14fc9d5406c46d3ff89ca9ff\",\"size\":9488,\"mode\":420},{\"path\":\"references/lark-whiteboard/connectors.md\",\"digest\":\"eed35832652927557318814d42f613afbf8d7d71e8a64c942084fda5d504949a\",\"size\":5734,\"mode\":420},{\"path\":\"references/lark-whiteboard/content.md\",\"digest\":\"a4c279552fc9f21f5049a1a07d42828594a8b1ca529d5afa8cbf7529d40e57db\",\"size\":1983,\"mode\":420},{\"path\":\"references/lark-whiteboard/image.md\",\"digest\":\"fe1ecd0048efecb5bca013322e79228cceca1cf3ec541def6af5fd2b75aa5c2c\",\"size\":4563,\"mode\":420},{\"path\":\"references/lark-whiteboard/lark-whiteboard-query.md\",\"digest\":\"289cbc32bb17b351956e385aa98f40ed2005bbc939dae5a25025b750cd3a7016\",\"size\":2268,\"mode\":420},{\"path\":\"references/lark-whiteboard/lark-whiteboard-update.md\",\"digest\":\"340b922e777305f3abeeaa58ec6c637f096a6c849c7b4ed5efb10b13b19287a2\",\"size\":3729,\"mode\":420},{\"path\":\"references/lark-whiteboard/layout.md\",\"digest\":\"cd2916a1d8bc00cba39ca8eed8ab3bf205cff6354517008e4c6c94a7d6b6eb51\",\"size\":20166,\"mode\":420},{\"path\":\"references/lark-whiteboard/schema.md\",\"digest\":\"02ec326b7e312780b7d72432b0d99cfd8b0e3b7d3d1529007462786ff965cc8d\",\"size\":15575,\"mode\":420},{\"path\":\"references/lark-whiteboard/style.md\",\"digest\":\"b1e17fc2990ff46af63c7b040a0a32a6c0ffca837166e9e9ab631f78415676e5\",\"size\":12176,\"mode\":420},{\"path\":\"references/lark-whiteboard/typography.md\",\"digest\":\"b78a0e16754be6f9e0d2239034214ed3264c7076dfe05a45e5e0cc81e74760c1\",\"size\":2535,\"mode\":420},{\"path\":\"references/lark-wiki.md\",\"digest\":\"97830f971f77bdfac194ca85f7b304471d1f313bd564ce09cda5f8b8dff4100e\",\"size\":7091,\"mode\":420},{\"path\":\"references/lark-wiki/lark-wiki-delete-space.md\",\"digest\":\"b4217cb81641d49e87dccaf93b137be86c7412690949eaf2f32ab06a2fee65e2\",\"size\":8680,\"mode\":420},{\"path\":\"references/lark-wiki/lark-wiki-move.md\",\"digest\":\"dcc93217ccd34c76e750d3cf3d5b4af06aa0eb78bfbef1f6e8576fcd060d92f2\",\"size\":10270,\"mode\":420},{\"path\":\"references/lark-wiki/lark-wiki-node-create.md\",\"digest\":\"ded5acfb262525d6a8aadc80e293fc37acdf799c3528a6bc6c1ddb509ff189cf\",\"size\":7176,\"mode\":420},{\"path\":\"references/lark-workflow-meeting-summary.md\",\"digest\":\"b7adbbd53ace2f639a01a8683733665c5b3311dbadd150b6f7b37e9780354b9c\",\"size\":4125,\"mode\":420},{\"path\":\"references/lark-workflow-standup-report.md\",\"digest\":\"668337ff574177747d7e255586d975568e793d268e2fb3dd330036b31742f202\",\"size\":4456,\"mode\":420}],\"digest\":\"1e0ed8834390381e680cfcce513003d52c193938fe07e188a4d8557cbb3c5039\"},{\"ref\":\"builtin:python-script\",\"api_id\":\"builtin-python-script\",\"name\":\"python-script\",\"description\":\"Create robust Python automation with full logging and safety checks. Use when tasks need complex data processing, authenticated API work, conditional file operations, or error handling beyond simple shell commands.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{},\"root\":\"system/python-script\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"f900fc35795fc1018fe086a5cd2b9d7682511960274b5700b2c4774c676266fb\",\"size\":1299,\"mode\":420},{\"path\":\"references/script-template.md\",\"digest\":\"7e681d0db5c956347d7d5e7b0380725beaf536e32375fcf79b8dac4e70a8ff02\",\"size\":1060,\"mode\":420}],\"digest\":\"7837df183e21e2b0fd9fc1e37943b796db632a40ae7f01243566b6a8d9c812b3\"},{\"ref\":\"builtin:recally\",\"api_id\":\"builtin-recally\",\"name\":\"recally\",\"description\":\"Reading assistant for saving, organizing, and recalling web content. Use when the user says \\\"save this article\\\", \\\"read this link\\\", \\\"summarize this\\\", \\\"check my feeds\\\", \\\"add to my library\\\", or asks about previously saved content. Handles articles, tweets, YouTube videos, GitHub repos, PDFs, and RSS feeds. Articles are stored with their metadata and indexed for fast search.\\n\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"author\":\"CherryHQ/stella\",\"owner_plugin\":\"system/recally\",\"version\":\"1.0\"}},\"root\":\"system/recally\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"a42fb637f661a46f38bae3a4a9282830104dfd53e4bf08a9a5d087f853f20f7d\",\"size\":7195,\"mode\":420},{\"path\":\"references/rss-workflow.md\",\"digest\":\"5e84e29d4e29686c62c0034402aa6085eff17286a8e4c749aa90ab8820a12ed6\",\"size\":1686,\"mode\":420},{\"path\":\"references/save-workflow.md\",\"digest\":\"d8f38e42b647a3b654b0a31932da2034a9125ae4ffa6a12d535ba769ec1a3ee1\",\"size\":4911,\"mode\":420},{\"path\":\"references/twitter-workflow.md\",\"digest\":\"cfc415312a5a0bcc73c23e93cde52ee5002e846fad41798261b532650910801b\",\"size\":2265,\"mode\":420},{\"path\":\"references/website-workflow.md\",\"digest\":\"2789838d5fc5d6b33322b34fe8a3c4508851ada7a99593ae2783c5f1587713aa\",\"size\":2682,\"mode\":420}],\"digest\":\"ad2f4a3c9d6addb28c38676da6df82676484b351cce09dd5cf746817c610f3f5\"},{\"ref\":\"builtin:scheduler\",\"api_id\":\"builtin-scheduler\",\"name\":\"scheduler\",\"description\":\"Manage scheduled jobs. Use when the user wants to create, list, inspect, update, pause, resume, or remove recurring or one-time scheduled tasks. Handles cron schedules, interval-based (every), one-time (at) jobs, platform job templates, and scheduled workflow runs.\\n\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"author\":\"CherryHQ/stella\",\"owner_plugin\":\"system/scheduler\",\"version\":\"1.0\"}},\"root\":\"system/scheduler\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"972cf6aaa251cfea29732b9d3d234ecf6f2fed8b787e59145e4f5861dce719f0\",\"size\":4439,\"mode\":420}],\"digest\":\"199deb7e2ccbe1eb4a3323c61544a3e3450cf6d5e1689410c360356850d73de0\"},{\"ref\":\"builtin:skill-creator\",\"api_id\":\"builtin-skill-creator\",\"name\":\"skill-creator\",\"description\":\"Create new skills, modify and improve existing skills, and measure skill performance. Use when users want to create a skill from scratch, edit, or optimize an existing skill, run evals to test a skill, benchmark skill performance with variance analysis, or optimize a skill's description for better triggering accuracy.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{},\"root\":\"system/skill-creator\",\"files\":[{\"path\":\"LICENSE.txt\",\"digest\":\"58d1e17ffe5109a7ae296caafcadfdbe6a7d176f0bc4ab01e12a689b0499d8bd\",\"size\":11357,\"mode\":420},{\"path\":\"SKILL.md\",\"digest\":\"baa75a289d3573958f48d6c0494833875c1240f64c76b256a0c1803ee14bbb52\",\"size\":33282,\"mode\":420},{\"path\":\"agents/analyzer.md\",\"digest\":\"28ff495369f27d9e2c85c133c85b596e9cfb3b672252e95b8f5ca56cc6e012f5\",\"size\":10561,\"mode\":420},{\"path\":\"agents/comparator.md\",\"digest\":\"58a5e00788c709dfba4090292e5f351cb40e3fd70975feecc1a4257d2cfcccfa\",\"size\":7569,\"mode\":420},{\"path\":\"agents/grader.md\",\"digest\":\"7a9ba21f96d846eacb438b5d8d97ee6d2d74589395cef788b29f5140b72e9322\",\"size\":9064,\"mode\":420},{\"path\":\"assets/eval_review.html\",\"digest\":\"1f9ba52cfd4f9cf528a2f26acfe097f15c92f048b947f28e84555cd564a9db2f\",\"size\":7184,\"mode\":420},{\"path\":\"eval-viewer/generate_review.py\",\"digest\":\"fc9d1b9243fe5ab6012ebd579bd76d0035de1b79fd3b969de114defab26478fb\",\"size\":16365,\"mode\":420},{\"path\":\"eval-viewer/viewer.html\",\"digest\":\"8ba4b172fdc74f009d087a7af06b7c7c7aefb19da0bf9b14d96892c72aa5c315\",\"size\":45753,\"mode\":420},{\"path\":\"references/schemas.md\",\"digest\":\"467025efdc41dd3e4171d992a9f28adc2a9cba27c0bb45478598737553c9f86a\",\"size\":12172,\"mode\":420},{\"path\":\"scripts/__init__.py\",\"digest\":\"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855\",\"size\":0,\"mode\":420},{\"path\":\"scripts/aggregate_benchmark.py\",\"digest\":\"123ef128ea5ccc01a4b1ac212ef5567f21e9c13d3d240609780beeb3200c49aa\",\"size\":14386,\"mode\":493},{\"path\":\"scripts/generate_report.py\",\"digest\":\"13df7118a3c50c83c4c3250a606d5f2b20b25a3d44cbc392b3d669ec75281453\",\"size\":12847,\"mode\":493},{\"path\":\"scripts/improve_description.py\",\"digest\":\"87d864570220b699fac52da309d2d6efdb060647bfebc74f768128e646accf80\",\"size\":11116,\"mode\":493},{\"path\":\"scripts/package_skill.py\",\"digest\":\"1a33059b0db1ef73375d46d513e5ea81369d2e8838c970597b0d52ddef8d1c0f\",\"size\":4234,\"mode\":493},{\"path\":\"scripts/quick_validate.py\",\"digest\":\"67cf5703402013936c8fb75ad6a1afecd8841d45cc5e606b634eb05825fde365\",\"size\":3972,\"mode\":493},{\"path\":\"scripts/run_eval.py\",\"digest\":\"43e3b8f80dbf69c343967ba77e268fae991d9fa3ed68b32a0ff02532cd48657f\",\"size\":11464,\"mode\":493},{\"path\":\"scripts/run_loop.py\",\"digest\":\"7bd6f674203168520517eec94c55f493c0d154339b061b4d7c0f0dad187d0f21\",\"size\":13605,\"mode\":493},{\"path\":\"scripts/utils.py\",\"digest\":\"3af8ae62c40c73ab712207436a0d9a981e845f25c5a7040229eb189cc8e45bb1\",\"size\":1661,\"mode\":420}],\"digest\":\"414822d4852b4511d8be6a799f58fc10712f370a83941844d8097dfb863a03ee\"},{\"ref\":\"builtin:stella\",\"api_id\":\"builtin-stella\",\"name\":\"stella\",\"description\":\"Self-knowledge about stella, the self-hosted AI assistant. Use when the user asks about stella itself: configuration, setup, onboarding, providers, models, agents, channels (Telegram/Discord/QQ/Feishu/DingTalk/WeChat), webhooks, memory system (LCM), scheduled jobs, reusable workflows, goals (objectives that converge through acceptance), workers/decomposition/dependencies, skills, plugins, session compaction, notifications, self-update, multi-agent, multi-user, or general \\\"how does stella work\\\" / \\\"help me get started\\\" questions. Also triggers on \\\"change my model\\\", \\\"set up telegram\\\", \\\"set up discord\\\", \\\"set up dingtalk\\\", \\\"set up wechat\\\", \\\"set up webhook\\\", \\\"configure provider\\\", \\\"update stella\\\", \\\"what can you do\\\", \\\"how do I install skills\\\", \\\"stella onboard\\\", \\\"switch agent\\\". Also triggers when the user wants to report a bug or file a GitHub issue about stella: \\\"report this bug\\\", \\\"create an issue for this\\\", \\\"报告这个 issue\\\", \\\"帮我建个 issue\\\". Also triggers on goal-model work — goal status, \\\"why is this blocked\\\", \\\"decompose this\\\", \\\"review/accept this\\\", \\\"为什么卡住了\\\", \\\"拆解成子目标\\\", \\\"验收\\\", or being dispatched as a goal worker: read references/goals.md. Merely creating a goal does not — write a clear intent and call the goal tool.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{},\"root\":\"system/stella\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"095069f204ab37ff272f7e8931a1598d7030b7fc93c8385aadb1bc8d6f39533b\",\"size\":28088,\"mode\":420},{\"path\":\"references/channels.md\",\"digest\":\"74137d86ca1d23e6d279ed8512d29649b9b0720994dfd69bf9cdb27df5167cd0\",\"size\":12135,\"mode\":420},{\"path\":\"references/configuration.md\",\"digest\":\"a4a1a31e7f7bfdb1dccab4f2c4bc531fc8cda3fdc568af34bd2891cb5f6cfb67\",\"size\":29273,\"mode\":420},{\"path\":\"references/goals.md\",\"digest\":\"056072761f3433b4005bf7aa912923c2d9883e28b90eb22af056ef990785654a\",\"size\":8885,\"mode\":420},{\"path\":\"references/models.md\",\"digest\":\"0734e738887b92e2dd8949f1132ce579bfd45ed53b747aef82994ecee88f6800\",\"size\":3001,\"mode\":420},{\"path\":\"references/report-issue.md\",\"digest\":\"dcc465955a7e95dadb3cf6bc8948f6c3c75bee24781719c19bebb2cd6a05be51\",\"size\":3347,\"mode\":420},{\"path\":\"references/update.md\",\"digest\":\"b3df654900385845fac46af4b01fafe685766af6336a749cf6cb2e96e87b660f\",\"size\":3700,\"mode\":420},{\"path\":\"references/webhooks.md\",\"digest\":\"f21d41006e7a1d5c4723f93535e416756751f475322e6ee745d71188bcdddbc3\",\"size\":1906,\"mode\":420}],\"digest\":\"9f61897dc2d1eda65540cc54cabd97015084a4ba36fb5ea065b6f6e16181bcf1\"},{\"ref\":\"builtin:web\",\"api_id\":\"builtin-web\",\"name\":\"web\",\"description\":\"Read the public web: search for sources, read one page as Markdown, and run site scripts for structured records (X/Twitter via FxEmbed, Exa search, GitHub, Hacker News, Reddit, Wikipedia, Bilibili bundled; a catalog of more installable with one command). Use for any task that needs something from a public website: a topic without a URL, a URL to read, a site's records (a tweet, a timeline, a repo's stats, a front page, a ranking), or a page that only renders with JavaScript.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"author\":\"vaayne/tap\",\"version\":\"2.0.0\"}},\"root\":\"system/web\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"41ccaff3e894dcbee50f2cf4131aef06750531a12d3f4e3260181fbac0e0e13b\",\"size\":11278,\"mode\":420},{\"path\":\"scripts/package.json\",\"digest\":\"72163e64f1d71a6d466645d5490a902e42a75d96118f86b3b2f073a55ff8b8a3\",\"size\":149,\"mode\":420},{\"path\":\"scripts/site.py\",\"digest\":\"46350ec43f5dd753f57b93b47ef66c270e3cf12fb414421254d011aa628b7cb6\",\"size\":16060,\"mode\":420},{\"path\":\"scripts/web.ts\",\"digest\":\"614dac68d1b7b9193c36d997f7f385c7bef49ddd070ede110a7933d1a65513b5\",\"size\":26662,\"mode\":420},{\"path\":\"sites/bilibili/video.js\",\"digest\":\"f36fefce96dabc9ee0b1c42f24f52cf09f7d468356985ab1ddc7e910236f49c4\",\"size\":2321,\"mode\":420},{\"path\":\"sites/exa/search.js\",\"digest\":\"e08ca9ffc6311fd96bd35bfe7d6567860c20a668e60d6f3d4e0da2335dec50a1\",\"size\":3396,\"mode\":420},{\"path\":\"sites/github/repo.js\",\"digest\":\"16fbd3d0bf103d63d087afe6dd50ddeea7b86e8754ace7ae5846762c633ddb59\",\"size\":2234,\"mode\":420},{\"path\":\"sites/hackernews/top.js\",\"digest\":\"68fb21108eda4e7d4cd012c002b08c3b5b764a7ca66289512605bb7d721ac49e\",\"size\":3799,\"mode\":420},{\"path\":\"sites/reddit/search.js\",\"digest\":\"8cfdd1688706fbbc570315e48974b589b3c928975106340c95d3faf7df4944f9\",\"size\":4243,\"mode\":420},{\"path\":\"sites/twitter/fxembed-profile-articles.js\",\"digest\":\"89a0c518360b845def92609c76b474d96c2066eae1477557c4f09e7f8e744eb5\",\"size\":1340,\"mode\":420},{\"path\":\"sites/twitter/fxembed-profile-statuses.js\",\"digest\":\"da3fcfb2ff1b972041e75497d780b40b6cbd5668d5a25ecb03521c33aaf1576b\",\"size\":1973,\"mode\":420},{\"path\":\"sites/twitter/fxembed-status.js\",\"digest\":\"64e030d0dd4df7e1b176b16217d5a14300045343cb679c4bc9b830c36b659efb\",\"size\":3529,\"mode\":420},{\"path\":\"sites/wikipedia/summary.js\",\"digest\":\"ea54081ec7c031c075513f62e7a9e67f89448486c2b70334391f1851bb0f889c\",\"size\":891,\"mode\":420}],\"digest\":\"4c71e38027590e6b77789b08881f9e8551fc4140dd43937f82fa72add039a961\"},{\"ref\":\"builtin:xberg\",\"api_id\":\"builtin-xberg\",\"name\":\"xberg\",\"description\":\"Extract text, tables, metadata, and images from 97+ document formats (PDF, Office, images, HTML, email, archives, academic) using Xberg CLI.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"license\":\"MIT\",\"metadata\":{\"author\":\"xberg-io\",\"repository\":\"https://github.com/xberg-io/xberg\",\"version\":\"1.0.14\"}},\"root\":\"system/xberg\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"110b3822dd43783022eb40a0aa7ff91e5a58eadc64ef7b6f95bb2a097f870c7e\",\"size\":4161,\"mode\":420},{\"path\":\"references/supported-formats.md\",\"digest\":\"8712ba5e292662a875b666574d81542d5527c4843e7d2bbba3e05b5092090983\",\"size\":17799,\"mode\":420}],\"digest\":\"f7e4268303b9e8f4633301aea3b3895399e3c61b239e4154439e211e96331b96\"}]}"
+const builtinManifestJSON = "{\"revision\":\"dea5330dc7b653139daee74a844ea2ff2331ad8ddfcd291ff2b6de288d93bd5f\",\"skills\":[{\"ref\":\"builtin:email\",\"api_id\":\"builtin-email\",\"name\":\"email\",\"description\":\"Standard IMAP/SMTP email client for Gmail, Outlook, and self-hosted accounts.\\nUse when user mentions email, inbox, send email, check mail, read message,\\ncompose, reply, set up email, add email account, configure email\\n(non-Lark/Feishu accounts).\\n\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"author\":\"CherryHQ/stella\",\"owner_plugin\":\"system/email\",\"version\":\"1.0\"}},\"root\":\"system/email\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"4df3c9a5f9f858c23745f66b95de788beb56994402a4164fc330817e09c021c4\",\"size\":3192,\"mode\":420}],\"digest\":\"bbd0a4528f9a59eb876ccc178571e07a0906c087d53e9bac2a4bbe05ce6c25cb\"},{\"ref\":\"builtin:html-artifact\",\"api_id\":\"builtin-html-artifact\",\"name\":\"html-artifact\",\"description\":\"Create polished single-file HTML artifacts for plans, reports, explainers, code reviews, lightweight dashboards, prototypes, and one-off editing tools. Use when the user asks for a standalone HTML file, single HTML, HTML artifact, interactive report, shareable local artifact, Alpine.js artifact, custom CSS artifact, copy/export UI, or a browser-openable deliverable instead of Markdown.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{},\"root\":\"system/html-artifact\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"03d360305bbcd78a6e50e934e06981dd54ab2d6cb66594884d5612e562f8e1fc\",\"size\":3284,\"mode\":420}],\"digest\":\"45d1e904378dee115664002fb8c3bfa96c84a5ad39c29e7b95ad08d89419ecc0\"},{\"ref\":\"builtin:lark-cli\",\"api_id\":\"builtin-lark-cli\",\"name\":\"lark-cli\",\"description\":\"Lark/Feishu CLI skills for Stella sessions. Covers workspace operations — calendar,\\ndocs, tasks, mail, and messenger — via the Stella-managed lark-cli tool. Stella installs\\nthe binary and injects a managed per-user OAuth token; no native CLI bootstrap is\\nrequired. Always\\nread lark-shared first for identity selection and permission-denied handling.\\n\",\"tags\":[\"lark\",\"feishu\",\"workspace\"],\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"generated\":true,\"owner_plugin\":\"tool/lark-cli\",\"source_ref\":\"2e4cfb49219d\",\"source_repo\":\"larksuite/cli\"}},\"root\":\"system/lark-cli\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"54ea27ac48fdb76072c2e2d1b6066d26b3add41f8caf5ddfa4addbe5ad1409a0\",\"size\":25555,\"mode\":420},{\"path\":\"references/lark-approval.md\",\"digest\":\"c03b4f65977087d63d7720768592a2f834ac6a21767a4d27572205fd3ae3e591\",\"size\":1639,\"mode\":420},{\"path\":\"references/lark-attendance.md\",\"digest\":\"776b012dd3d6bc50e510cc39ef416e94b636523f88af26f24a12ec7eaa00e4a1\",\"size\":1720,\"mode\":420},{\"path\":\"references/lark-base.md\",\"digest\":\"c95ea3c9b14d9f910cd711286bcd13651fae3dd60c58ce49e2bc85dca9072008\",\"size\":55109,\"mode\":420},{\"path\":\"references/lark-base/dashboard-block-data-config.md\",\"digest\":\"0f55f98760e76ef8c0eecfa9cfb0198c9a3f56be4387b910ac8c4640e0bdce00\",\"size\":15182,\"mode\":420},{\"path\":\"references/lark-base/examples.md\",\"digest\":\"49c4b3f1ab3770a0a2f6e82d9f6885f9310d40e0c39ef3dfe19c0e07172f2461\",\"size\":3637,\"mode\":420},{\"path\":\"references/lark-base/formula-field-guide.md\",\"digest\":\"72b14241bbb55c7cf50481c689a45fcd4462c13bf651031b689f740bdaecbf7d\",\"size\":56980,\"mode\":420},{\"path\":\"references/lark-base/lark-base-advperm-disable.md\",\"digest\":\"88f6f51163d6bae21bf5b5d681fbe0194cf73eed14263f03bdb798b24914be5b\",\"size\":2485,\"mode\":420},{\"path\":\"references/lark-base/lark-base-advperm-enable.md\",\"digest\":\"d73c18457807ae7f08ea5762f60f6659f2fa7328103ccb78a07d90f96d95b56d\",\"size\":2307,\"mode\":420},{\"path\":\"references/lark-base/lark-base-base-copy.md\",\"digest\":\"5cc2fe1478a299402345411d5d4c1cdfa6fa46fec4e5bd12d34178bd6f9b1467\",\"size\":2938,\"mode\":420},{\"path\":\"references/lark-base/lark-base-base-create.md\",\"digest\":\"89ff76f094acf6a75793726f3025ccbd8850cc3a36a7cc5e2b61b5bcc1a904b4\",\"size\":2606,\"mode\":420},{\"path\":\"references/lark-base/lark-base-base-get.md\",\"digest\":\"7bd335528ac4f0a312bd13667fd7836ea2b25accc3b5b65f16fdd5b0c6d44e3f\",\"size\":1025,\"mode\":420},{\"path\":\"references/lark-base/lark-base-cell-value.md\",\"digest\":\"7a3f9bfe433de84fa040b9e4a72974494d84490ee49eff35b49a290996ae0747\",\"size\":4152,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-arrange.md\",\"digest\":\"2edf273452d14af35abecc0b75d3820f6ad9231c0abe74de91f903cd1b9c6d97\",\"size\":2816,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-create.md\",\"digest\":\"4892f407b5ddb1967c1aa2d9dbe2f577a9d1ad898c3a1d416b092d02b06823a1\",\"size\":5374,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-delete.md\",\"digest\":\"685539ea10fd53862aedf0cf84a03771f3fec37738d841265744ed99b1324977\",\"size\":1177,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-get.md\",\"digest\":\"cad9aee28924775191eafb605ac74439c808d2374cf5eb9dd2c7d22013f75ba9\",\"size\":2181,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-list.md\",\"digest\":\"e4cdf98d4809dc0e4f460ed753cc61c5ab5849793841345031f0b4ce168a20fe\",\"size\":2051,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-block-update.md\",\"digest\":\"c2d62a867e667e19557d63fa7bfb1222e4546475956f5b8ec6848c16411fdd68\",\"size\":4444,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-create.md\",\"digest\":\"e6b5c0b6492fa6f96bdf0be3839fa2eb3ecde7294253b8409fb9d029f5954b6d\",\"size\":2255,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-delete.md\",\"digest\":\"79973846ea4afc558c349256d51ea1896352bed91bc73dd78fcfb2b85fe4af3c\",\"size\":1191,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-get.md\",\"digest\":\"071e6b2d1be3afbb08dc71bb27bb1bfc717bc1c937a24918b3bf9c39be3ace45\",\"size\":2331,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-list.md\",\"digest\":\"0b83f66ef8615d8d65b6c3d3d362a00ab806bc6375d54b7838b4d8bec7252922\",\"size\":1980,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard-update.md\",\"digest\":\"2d4b5f0e63f3e48b7a69ecdcb4d0a96fa1cc5696b1aed0bfdd99cb24f3cc0896\",\"size\":1924,\"mode\":420},{\"path\":\"references/lark-base/lark-base-dashboard.md\",\"digest\":\"2c76cdf7286878885c9b05d80fdf4b2fa9f2a767448f263a9e892dabee052ecd\",\"size\":12588,\"mode\":420},{\"path\":\"references/lark-base/lark-base-data-query.md\",\"digest\":\"c4004baf4533448f696a57a114a0f8dd6c95fa2f7625745fbb46cfdb550ae054\",\"size\":19050,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-create.md\",\"digest\":\"773fd641f49edcfe4c5e1a5b01baaa0e1bee2757176dbf0b8aaa4a4e4969302c\",\"size\":3578,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-delete.md\",\"digest\":\"bdabbaa205ece1016c857e81187825ac056bec8cbd482ab23f02198da024b64b\",\"size\":1407,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-get.md\",\"digest\":\"320cc32a1e4286ecd0595161aff65742eb4e9a449449a57c83e208ff3d215004\",\"size\":989,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-list.md\",\"digest\":\"84f581aa1afe0bd596bc5b5ad6691d2d7b5bfb80ca3d121153d4251b424d2aa2\",\"size\":1190,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-search-options.md\",\"digest\":\"e1b642295d13ced7ea5b28b80f006eee3119cfe34f39e18b11e760cb97749246\",\"size\":1489,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field-update.md\",\"digest\":\"3c602ca9b956e640c08f4f48c6b34451e3a20467ee5d35a18b1d4101e8d18651\",\"size\":3463,\"mode\":420},{\"path\":\"references/lark-base/lark-base-field.md\",\"digest\":\"099a943d857ebf8acfaf6d5068b46f8513e73d8f63a0a0b7f566473275bf0225\",\"size\":1554,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-create.md\",\"digest\":\"23ba0d15cab80c788a53d75cf59eb84bd10860c3802bb483020dfac22a0bb8f3\",\"size\":3092,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-delete.md\",\"digest\":\"801058bf1aa463ac0a357562eebb06c4bc3d47f45bc8b6b6653867844f7db486\",\"size\":1792,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-get.md\",\"digest\":\"2c8a267ceca1d400cb0e0707dd5a44c3fdc28ef491c151f901dd87bcf182c98c\",\"size\":2084,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-list.md\",\"digest\":\"7d3df88a25e96ea7d8c15dc31fc2fd4966ec5546e508c9d832255ea48bbc52b1\",\"size\":2480,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions-create.md\",\"digest\":\"7d0a4408bd2152de30aa396976c494d68bd7e75fe3a2d7dc55cf22a3777fbc10\",\"size\":6411,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions-delete.md\",\"digest\":\"c5ea081bdcf5ccb317628b4e43a0bd205a0347e5488cc9d19f9a6968ce5fb6b1\",\"size\":2298,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions-list.md\",\"digest\":\"361932c6c07b1be6f189cb453f15a826f7ef824e925f3abaa77dfe3f063d79ca\",\"size\":2570,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions-update.md\",\"digest\":\"212e2aa37cdce5d49d9444acc6903b3ba6effd3ac51d4fd650dbaa1cf735261a\",\"size\":3708,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-questions.md\",\"digest\":\"8ad00fe11861df9562b01a9918d9ffeaeee94b79b645bb7da02db411b8a875d0\",\"size\":1254,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form-update.md\",\"digest\":\"a1f07ddb86ee24708a6fca8fa7ac3d4ba5df0d9a84943e8775f20c073b3000b4\",\"size\":2807,\"mode\":420},{\"path\":\"references/lark-base/lark-base-form.md\",\"digest\":\"bcc1b20cdc5c962805c61b187e0c317d379a6eb93e89944a1d553486855bcdaf\",\"size\":1187,\"mode\":420},{\"path\":\"references/lark-base/lark-base-history.md\",\"digest\":\"d0b1f46cbd7c837512b1fe6ad931195689bc15cee1a811561646595f1c54fcc0\",\"size\":870,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-batch-create.md\",\"digest\":\"c5879c14c0c589ebdef3ff59b3422dfb7bccc5a0be93d6c663b76f7eda2c11d6\",\"size\":2670,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-batch-update.md\",\"digest\":\"e39e53b0741466c0af0358666078b89cf8da35770df12933b1b698be91b7d987\",\"size\":2943,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-delete.md\",\"digest\":\"35749af2d51d47f95633bb83a08e46d641fe94927e1b726a620caa2537a42e19\",\"size\":1380,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-get.md\",\"digest\":\"b7bb7f7b10efe021736de5240ca50e419e71e44e3710b78a7dc7faecb2d51773\",\"size\":916,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-history-list.md\",\"digest\":\"10741120300d77dce7e382f514e888ed066d4939ddcb583cb641e6b75a4d9d1b\",\"size\":3249,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-list.md\",\"digest\":\"aabfbfec0dd7126397199b428dfdc0d14715561b5ae9b55d9d0488b1d7beb886\",\"size\":4286,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-search.md\",\"digest\":\"8d18000e59c11774dd95017356a4da1c72546cce4353ba8df0cec4667d3b2468\",\"size\":4263,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-share-link-create.md\",\"digest\":\"592c324f47df6a61084a8fe9ba327724a3e7aef37eb855d6cce0b216060eba32\",\"size\":2208,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-upload-attachment.md\",\"digest\":\"6f3e29aa9b8281405faff37a80f22c66dba72939c6061e973dbc805bb601f3f6\",\"size\":1959,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record-upsert.md\",\"digest\":\"41846e79fbb70ce966e7ce7c45c6d3efa3c1e022354c8bf3315ed08b4fccd8be\",\"size\":2666,\"mode\":420},{\"path\":\"references/lark-base/lark-base-record.md\",\"digest\":\"060d5074993435ceb5cb21b8f028fb5f7b4274c47262fe8493d5fcf955707d0b\",\"size\":3705,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-create.md\",\"digest\":\"966a82d2e9608686849626b72ff099e0facfec1f647f7aa64174c8dc3cda570a\",\"size\":3650,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-delete.md\",\"digest\":\"3b23ee86d43dceb281282411df36e15c5925d2442b4378a6f633420f0f060716\",\"size\":2387,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-get.md\",\"digest\":\"8851569583e1d8d61e8ebbd92d4f48091e1a6a0336fcd970af63882941448054\",\"size\":3307,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-list.md\",\"digest\":\"ca5bc190347b02307f18ee14424b50aee62cd6cc574886cfb8825927bcf029e3\",\"size\":2933,\"mode\":420},{\"path\":\"references/lark-base/lark-base-role-update.md\",\"digest\":\"aceae4ab238a47b1d73270f65bf5023a3e89cdf70236263266e03a6eaba56972\",\"size\":3166,\"mode\":420},{\"path\":\"references/lark-base/lark-base-shortcut-field-properties.md\",\"digest\":\"85abdc3999e38376d2ef5e3e8f8530fc57fa6b6486144b9d438a8823e05369e7\",\"size\":32026,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-create.md\",\"digest\":\"159704269dc92d584f57febef511a9a8eada96b6fe85eb94193383b451a442ad\",\"size\":1843,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-delete.md\",\"digest\":\"c101e61750e9312929cf8f830ae2b494df6e12965dd47685c4f9626822756679\",\"size\":1414,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-get.md\",\"digest\":\"74123d055adf7725244e48f5646b8ae26ffdcc04a87e743f3098ed0b0a8ad6a5\",\"size\":1480,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-list.md\",\"digest\":\"670624a861c3e66d0525713cb72d5fbe26a789bd883052df0388bfd2ef961d99\",\"size\":1103,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table-update.md\",\"digest\":\"7b4115971b9aaa88be8cfeb068e4ad0df91f4fa60de8ab9355810fa89b636c97\",\"size\":1088,\"mode\":420},{\"path\":\"references/lark-base/lark-base-table.md\",\"digest\":\"e000075679e5d8b9d51e2908c26d34d4e9aa4c89c14f788253196debdc19365d\",\"size\":1277,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-create.md\",\"digest\":\"81dddf4bffd7d805c277cbb8e2d8d3f51247bd20ae7c987bf4a1b8b2515c5a29\",\"size\":2582,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-delete.md\",\"digest\":\"6544b0beb80098cf16487e789050815b1af4a7d6e8ece9db8da37e322788ec67\",\"size\":1061,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-card.md\",\"digest\":\"215970e06056af1b73cd34886fa8ab4ace8991e14308bed142a010b8f8999a78\",\"size\":877,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-filter.md\",\"digest\":\"bcbd9e8df79ce9d24e562d09fc3ec513da6738d890fdcd525c45a0f318290038\",\"size\":883,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-group.md\",\"digest\":\"24a005d1ad8bc61efd7b8de5627546388787b7b31a23509f7be6d7f6a7b4e51c\",\"size\":880,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-sort.md\",\"digest\":\"1bfb13dacf7ecb9305535ccf198bfe844c583199544c7458499b5ecefc729be4\",\"size\":877,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-timebar.md\",\"digest\":\"aec83ffd1c7126304553e358bb4f06d62dc2fe3c422336102067c7e535f138d7\",\"size\":892,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get-visible-fields.md\",\"digest\":\"e29d32230c791c2f21bec96b8e145d0426219fa5669431848ec0259dc68c8cbe\",\"size\":922,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-get.md\",\"digest\":\"1a7866e35f66e9a29447cfcbbb42b8127f311471c5106a5ab32af4dd5e20b33d\",\"size\":880,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-list.md\",\"digest\":\"791157d8afa07361b03f6cc79e4737d31e2bb763824d363b726abc9e42c0aac0\",\"size\":1183,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-rename.md\",\"digest\":\"a048be3869d620118c27347b6b1b0d35300fa5a08a4625509cd5b906f3ae1d39\",\"size\":1038,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-card.md\",\"digest\":\"0ffb20fcdae884503c9940ccbc3f6556efd70804d11e12da1924df9fe7ce4dca\",\"size\":2251,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-filter.md\",\"digest\":\"4dd7804ce91772294169072c20d3fd2511baad6664ef385ec4370402de13b213\",\"size\":6534,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-group.md\",\"digest\":\"42366cbf8cbf1e2e3a048b6e569880734ca14ba63bb69c052cbcf53e53acde14\",\"size\":2403,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-sort.md\",\"digest\":\"ab7e1ad8070c28b1fdeaf827eee59d73584f4973705912d61345ed3e052beba9\",\"size\":2291,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-timebar.md\",\"digest\":\"704528a273a50e4d93013ecb5a6e02f47cd2e1c1b7ca5f8e66d05748652b3ee9\",\"size\":2736,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view-set-visible-fields.md\",\"digest\":\"c50d32e6dce25f4a71fbb67eda7dcbfde72bd3d90b3c80ebe26321b0f1fe50d0\",\"size\":2180,\"mode\":420},{\"path\":\"references/lark-base/lark-base-view.md\",\"digest\":\"32402cd5083e8b37575a855d9e9fe7e6fb5dc6da9b13e8479d49f211b3115155\",\"size\":3453,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-create.md\",\"digest\":\"ab54d32128bfd6a89319485d2f4c25663606b9a6d31740965e9f693ab9f6b534\",\"size\":8878,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-disable.md\",\"digest\":\"911ea146d1fee02ab1d4fd53080874dc0b94b38a969b9356e9598466362f398f\",\"size\":3156,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-enable.md\",\"digest\":\"d499ae0b703875fec0406d8f11775399a3c3b13d0bc9003c72c76d469dc73641\",\"size\":3148,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-get.md\",\"digest\":\"b926f557d9aa019bfeb30b186ca631c8a58b4042a11d5370788bd6059226cc14\",\"size\":7464,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-guide.md\",\"digest\":\"5fa91e39b21e3c3f2a32b96c0eb0edcddc64f9fe903b318615472986959fdb47\",\"size\":26427,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-list.md\",\"digest\":\"f0433cf7dc40dd8eb3bbbfee994e760ffd017e302031818120383d35edfcadef\",\"size\":5212,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-schema.md\",\"digest\":\"89d8610ad1ee3312efae8524112e9c7ac6613acc2c2a802c34376627ee823081\",\"size\":49139,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow-update.md\",\"digest\":\"7bbb8baebc4bfcd81f34df5935bb8b757e94005d2f1c5597379dd2b58a4d5cd3\",\"size\":7553,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workflow.md\",\"digest\":\"11bf335fc68bef918a79664e9185a1694c29556315fc9531ba308fdf0ae3a9ee\",\"size\":1997,\"mode\":420},{\"path\":\"references/lark-base/lark-base-workspace.md\",\"digest\":\"812be3c7ace8bfed55b615e4bbd8516b464c76fb6405479a454446bde6bbf0f2\",\"size\":845,\"mode\":420},{\"path\":\"references/lark-base/lookup-field-guide.md\",\"digest\":\"6382f4ed07b01a851243814854b29a98b886a052c68a840407ddcf97d4a59e3d\",\"size\":21979,\"mode\":420},{\"path\":\"references/lark-base/role-config.md\",\"digest\":\"19316f0a26a87706a7614df1b565880f00d857c561c666ab8a0dd53943ef7cd3\",\"size\":25503,\"mode\":420},{\"path\":\"references/lark-calendar.md\",\"digest\":\"ad558e85f381c2a20f8f8d602b17264af0e1f441348bbbacdcf3afe7d31c7561\",\"size\":15421,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-agenda.md\",\"digest\":\"c2054ce3c43ecbe88ea44403e9e164f7f1c7d9b2ddb1800abbd14f3a09bdac04\",\"size\":2848,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-create.md\",\"digest\":\"c6240fafc05bfc073f6ceda6d978acc98142bbb483f4f5f7c6a682e756f6691a\",\"size\":6333,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-freebusy.md\",\"digest\":\"810a6c2c005ca45e9122fca5704c5cabf251e50f88c6b6afb6361d677a0d1b0c\",\"size\":4617,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-room-find.md\",\"digest\":\"fdff4b26e8b8e4dbe33272389758cdc497552731ce783bd2b3f1d2d4fa285fad\",\"size\":12490,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-rsvp.md\",\"digest\":\"0c7e09fb2f1d07bc1bb06e033689cc64b615c38e88ab5209ba301913a12c96e4\",\"size\":1726,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-schedule-meeting.md\",\"digest\":\"4773c4fb81859d0643c3f7802d9c02a2b57bedc16f9902ad551a4d8e782be117\",\"size\":19081,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-suggestion.md\",\"digest\":\"9cf2d32f4ee60b2fb814de1135972140e94a9bca618f593183c9b47346b5f25d\",\"size\":7910,\"mode\":420},{\"path\":\"references/lark-calendar/lark-calendar-update.md\",\"digest\":\"967c398d9d4816a5a555e5af01d27bb514c29aa8e77548b7fb02f72ed97944f0\",\"size\":8433,\"mode\":420},{\"path\":\"references/lark-contact.md\",\"digest\":\"7386d4a80c881447231dabe03de74a073fc06c38a8f37bf23ece1e8fcf1b255a\",\"size\":1322,\"mode\":420},{\"path\":\"references/lark-contact/lark-contact-get-user.md\",\"digest\":\"47c7190638d40fd5071b65689e2b99609d61b3d9ac1a302225401ed17deb38d6\",\"size\":1771,\"mode\":420},{\"path\":\"references/lark-contact/lark-contact-search-user.md\",\"digest\":\"072d092c96f87b48a3c4974202eae361d7c1ac10f6edb0b437806c987e0c0dfa\",\"size\":1499,\"mode\":420},{\"path\":\"references/lark-doc.md\",\"digest\":\"a43c74ed2b0248dcde9d0a16a780e1c7a90c2b80a5b872bfdb03940c575d787b\",\"size\":9199,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-create.md\",\"digest\":\"26360050ff601e0e6a0b2e24e2fd678bb955f0617606b7149d83e764f72f61c7\",\"size\":5464,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-fetch.md\",\"digest\":\"e23367bcd0795f5e14bb8a1c2d647bebc26ae39a84c762eff79411bd170b1226\",\"size\":9673,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-md.md\",\"digest\":\"e4108795236bdc0c8d1fabcb8652b521aed0f6fde5e5058473e969bc880a5837\",\"size\":5987,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-media-download.md\",\"digest\":\"58bc49188b05771fd247ae744a5529bd67d551c3baf45c4ae85e7df907629017\",\"size\":2350,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-media-insert.md\",\"digest\":\"649934b2a1458b51888b5c7fecfb7c8cd32069ecbeca51f0195b5ff9475ad91d\",\"size\":7525,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-media-preview.md\",\"digest\":\"27ac73168ef601c7ffd4eb4cd2a4d7c8f6ee1058d1498ae75e4613d175a3194f\",\"size\":1743,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-search.md\",\"digest\":\"7714586777209b57b55bcc1277bcc828fd5b5a11e456f28b444f109ee6029ef1\",\"size\":17975,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-update.md\",\"digest\":\"4b3357cca4940150deb5acc359fbf9e3d3e98d679537df450b29f4c716e30455\",\"size\":14595,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-whiteboard.md\",\"digest\":\"d9dd5922837a756ac9448e545c13253aa3d269e3a0215ec68bc683ffbad411ac\",\"size\":3717,\"mode\":420},{\"path\":\"references/lark-doc/lark-doc-xml.md\",\"digest\":\"01709b80873ea512f3aaa07bfde6aa7f2b360cc9fa6a2b6e07255151108286f3\",\"size\":9143,\"mode\":420},{\"path\":\"references/lark-drive.md\",\"digest\":\"4ac2f567cb351cd36c0ef93ca27b3a0bfa92fb162aa546419da1bfaab255e2f0\",\"size\":23520,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-add-comment.md\",\"digest\":\"4443d6b6918aad75a878ed8a0c30901d30d9f7459002c94aecc28abea5a90e21\",\"size\":9748,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-apply-permission.md\",\"digest\":\"f6ca8a77155bd45d9bb8d3ba5d66b9af0e4df0b7b33e16f03e627adc8cc53897\",\"size\":5461,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-create-folder.md\",\"digest\":\"584d64754b35926b993ae66585ed8f744f4b14466bb4270eaed5a450e15311d4\",\"size\":3745,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-create-shortcut.md\",\"digest\":\"9ff53f2e77fd96aa3eeed52934ad9997332047430cf803c464abb65c6e74f3f5\",\"size\":6090,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-delete.md\",\"digest\":\"2d6aa2b941eeb472a77bac7c3798e77ff285b507a773234e4228278d18785155\",\"size\":3047,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-download.md\",\"digest\":\"db665ac0a1210d48dd9c7d4bad018a36ba4860aae013dd01600be45d9f627bed\",\"size\":754,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-export-download.md\",\"digest\":\"5d0e13f82dd2a8ad91caa02f4222cb8469931479da65a99fab9c8270b2df8249\",\"size\":1616,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-export.md\",\"digest\":\"a4b0f8487b12c5891e616dc6e4371581950339a493e342f026922e37771413d2\",\"size\":3593,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-import.md\",\"digest\":\"3442fbe42141d7a34d85b2524858b3eea4d8bd84805d4ab7255c62385669d535\",\"size\":9876,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-move.md\",\"digest\":\"719c3c9f9889f0a69bcfbe1dd7f6fa013764c90f23fb0f03b0b3ff78ab9fdcd2\",\"size\":5547,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-reactions.md\",\"digest\":\"7c6761ac97e2e65f1ddb9a368a916a2a2b17b2d7c8b3146f1cad531b394c720d\",\"size\":6271,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-search.md\",\"digest\":\"a7449faa1e7bbff03082c0851a23e7e7ab3496b51e3d86076274e98b739df591\",\"size\":19496,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-task-result.md\",\"digest\":\"b17c30a6481924fdbbad8bee4a50da678b4ec9c708aa886d8e4b725af73c9ddd\",\"size\":12191,\"mode\":420},{\"path\":\"references/lark-drive/lark-drive-upload.md\",\"digest\":\"3fc4f91b080111e0198b59d2b76bb17529231a18cdb929215a068283958da0db\",\"size\":4784,\"mode\":420},{\"path\":\"references/lark-event.md\",\"digest\":\"5fcd0040bf6a95dcdc96072ece510703d5d7ae4ef5e78c88b7f297f68a2dbb31\",\"size\":9685,\"mode\":420},{\"path\":\"references/lark-event/lark-event-im.md\",\"digest\":\"4ef2de1f09e03f70c2549dc8de4ba3d72e6df6b4262b94eaca52666107e3835a\",\"size\":5480,\"mode\":420},{\"path\":\"references/lark-im.md\",\"digest\":\"124f0668b2f72c8c093997c325ad725ea62d408e54390950d6eaf390fcf080cb\",\"size\":12516,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-create.md\",\"digest\":\"b8bc93ed6050f57780c5cfdba89a22b03657f24dec398f8da6dc94e8235c21b2\",\"size\":11193,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-identity.md\",\"digest\":\"80d1d1ff90ed7c2583c8bd5d203eed0fd411d68f44f7d426cf2d07f727e84bfb\",\"size\":3596,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-messages-list.md\",\"digest\":\"09cdc14e510e5b76995ab8a09dd0d9958395e267246ffd2c9bfc88c08cab421c\",\"size\":12016,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-search.md\",\"digest\":\"243571538c313d09d7f740ab6410f1126584be7953b6150a8026f7d48cacc9b3\",\"size\":9770,\"mode\":420},{\"path\":\"references/lark-im/lark-im-chat-update.md\",\"digest\":\"607bf1293c5c0d56db22647a1b1f001461c635c7ebf1cc607ae550705f225aeb\",\"size\":4041,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-mget.md\",\"digest\":\"dab7948007922274ad6a0b6532571e5e52b663cc9c5d8943d31788a24f0a07dc\",\"size\":4445,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-reply.md\",\"digest\":\"a01d4430af97ef6099b674a9deaffa405d79ec872e96746ef26acee0c71667f2\",\"size\":13773,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-resources-download.md\",\"digest\":\"f75d8cdb67cd5772b27da7bcf9482da5be3ce12fac2daf34de4c7bdcd37fe357\",\"size\":7625,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-search.md\",\"digest\":\"a90a02bcc8cbb1aa0dcebf4baa6f54ec15c2dc478505a3ce3dc4fe3bf5a6dd63\",\"size\":14743,\"mode\":420},{\"path\":\"references/lark-im/lark-im-messages-send.md\",\"digest\":\"ac421c43664f90fe027d8c5fc210ab1e2ad8bdc219c42cb6edae14627f267aa2\",\"size\":15788,\"mode\":420},{\"path\":\"references/lark-im/lark-im-reactions.md\",\"digest\":\"b9af652a41a73af6a84acd83dddbdbab700a3b44655bbbe2024388da223fb106\",\"size\":12015,\"mode\":420},{\"path\":\"references/lark-im/lark-im-threads-messages-list.md\",\"digest\":\"89b9912f3ae856c6935ef51e954ae8444dbf2b4a9c62a8fcf02361bb2afa3651\",\"size\":5841,\"mode\":420},{\"path\":\"references/lark-mail.md\",\"digest\":\"11cb64efe6d11ebe8b9751d3e57a285f3e21d255f8a507b0695deb2e10540f0e\",\"size\":40944,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-decline-receipt.md\",\"digest\":\"267fe11794a9d41eaa848607f7268dd8cd550dbacd5789fd9f4538e2b7d7548c\",\"size\":4180,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-draft-create.md\",\"digest\":\"ffcafe78fd925afb97e2da61f8f8c5ac189bd944f5be07a04e28a2d2d6640a84\",\"size\":10075,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-draft-edit.md\",\"digest\":\"8bc23f56ccdefa2d6e0afd10978c157639ae14646bb2d5b1ebd0bdc2ffe6afca\",\"size\":19802,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-forward.md\",\"digest\":\"36033f1af3bf5b896463a740ef087cce53ccc02e5401a8eb51e17dd84fbd0a81\",\"size\":14531,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-message.md\",\"digest\":\"3cbfcf78446212acc1d49fcf5e02fa904797daca275c5c8fdd3410766f510e1e\",\"size\":13467,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-messages.md\",\"digest\":\"995fbb637dcdca97991a0520dd8310dbb84cd6ac22903679ee6d674991c7cd9b\",\"size\":5139,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-reply-all.md\",\"digest\":\"2d4032859fd58b6a3cd584211de89f61d0f57c3bdec7bf2394cd12a355ee622a\",\"size\":13885,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-reply.md\",\"digest\":\"f5baa3e0f44111237437533c96927143df5222452f0c9d2ea0f68a31ef2a9889\",\"size\":15390,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-send-receipt.md\",\"digest\":\"cdbb185cf4973d80f6186811d0b76b4cfcc4285fd6eb047c407ef9a603a08ad1\",\"size\":6424,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-send.md\",\"digest\":\"45a8862842fdf4a803a00845ac1c60499aadb22590815f732bb7b8b163407272\",\"size\":14270,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-share-to-chat.md\",\"digest\":\"ea94601c4c747ef1cb9dca8182e73a3a82377a28124d6e75f4125025fbc5e8c3\",\"size\":4537,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-signature.md\",\"digest\":\"38c0378777fe466cf9406108d2f00db3015b623e55bfc23c15cb3d344cca4928\",\"size\":3199,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-thread.md\",\"digest\":\"dff999f72264c64e43d93f55e67cfc6da9089fd3546ccc8cfc83889bed4749db\",\"size\":4868,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-triage.md\",\"digest\":\"858ef54e2d80936aa72b741af94d4b1cb6c74f3b2f8b3a0ae540ef834601b121\",\"size\":8290,\"mode\":420},{\"path\":\"references/lark-mail/lark-mail-watch.md\",\"digest\":\"b30ac05ff2f9ce8b807006e9744c8e4e38121c1ab0d38004faa3f1f8e9a96937\",\"size\":5681,\"mode\":420},{\"path\":\"references/lark-minutes.md\",\"digest\":\"d92e5f73a2f15380a4611e5623f9163e00112833dfaeba69db2fa7e9e63f8eaa\",\"size\":7401,\"mode\":420},{\"path\":\"references/lark-minutes/lark-minutes-download.md\",\"digest\":\"80a2aba753bb5af47673c8b8b71290e1e57f6e79e3623603241a0962f3ddbb54\",\"size\":7470,\"mode\":420},{\"path\":\"references/lark-minutes/lark-minutes-search.md\",\"digest\":\"1f0fd7568ba574b6a7b2a499cfd25a7c42925e6808c71b8705507500525a4aaa\",\"size\":11683,\"mode\":420},{\"path\":\"references/lark-okr.md\",\"digest\":\"d33520bcc20579556a19ca5bcb2cc1fbeb80638af38f1560198d92d4148b4b48\",\"size\":6292,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-contentblock.md\",\"digest\":\"f65a3b210ec1b3ac7c2aba359a3b6692f4c1c87335379e76772f920284105fe8\",\"size\":9604,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-cycle-detail.md\",\"digest\":\"86a3dc4018748426def1ca118a9eed6321de77ea4be458951007b7f9fa6352ab\",\"size\":2560,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-cycle-list.md\",\"digest\":\"de492e806d6bfa23b562d90f39aac9a43cf3b93d2651f227e9f3680b6d58134e\",\"size\":3882,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-entities.md\",\"digest\":\"a51d7f4f85b0ed6ba36e95b04af45aab7f74fa0cec87e84971beed8b06086a09\",\"size\":21859,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-image-upload.md\",\"digest\":\"777aea7c0d562fc2014f9dca28b4f910f1a58f0e9dee487a19eb324ad31276f8\",\"size\":3554,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-create.md\",\"digest\":\"811a78881300565c41287bd3847addfb9c67064024b26d1269f3455a759e689b\",\"size\":5836,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-delete.md\",\"digest\":\"a0a4b9f124a78f29b8e505a842abc01d8194bd6dceb8add0a5cbdd8cece8f4bb\",\"size\":1489,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-get.md\",\"digest\":\"f849c4a8f3b3abb213dde6da83019ee218ec5d40e16c9cf7dc00c6421efabd02\",\"size\":2186,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-list.md\",\"digest\":\"ee1a2ea7df138fb662b8769c9a2792c95561d43b71ca57f275777a2b2a6ccbee\",\"size\":3282,\"mode\":420},{\"path\":\"references/lark-okr/lark-okr-progress-update.md\",\"digest\":\"72d8f99dbf95c615d72cef192fec9522b4933881290e336504d1d7282d4dbf8f\",\"size\":4966,\"mode\":420},{\"path\":\"references/lark-openapi-explorer.md\",\"digest\":\"822cf3dd71610824a640e8bf67354c22b94cfc78ec701c648287cb2dc7290b2c\",\"size\":4982,\"mode\":420},{\"path\":\"references/lark-shared.md\",\"digest\":\"13657eb5c89ac6e5c716cfd9ea1e8c0570abb2edc332fbc53b26ff38a7bf96b2\",\"size\":4242,\"mode\":420},{\"path\":\"references/lark-sheets.md\",\"digest\":\"e3a8854db2d016049de9bb71916197a992fde724a96480c1a8b5e75816868054\",\"size\":20152,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-add-dimension.md\",\"digest\":\"0c46243beb929a7358d52b080bff434925385daeb061d5a20950f7d84bad0d2b\",\"size\":2077,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-append.md\",\"digest\":\"f92479bea20b8ab168f3d0474d7dee3c64b95dc75555047eb809bf3acfebfd80\",\"size\":2497,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-batch-set-style.md\",\"digest\":\"4645b41ea31184fc5dc59de20a34a5924813c34dc65a5225750845fdeab0fdc2\",\"size\":2324,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-create-filter-view-condition.md\",\"digest\":\"0506659b45afe666bd0d3415e653489377a5b0094c9df8c2a1c21f7f2e718eb9\",\"size\":1970,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-create-filter-view.md\",\"digest\":\"0e408421cab5eecf2aa04e39e479b8d6f6bc037105e6e1f7699f8872342120c9\",\"size\":1826,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-create-float-image.md\",\"digest\":\"0419d50105b5816dfbdf737a878850bbf5c4d5cc1e10611ca9d614ebcff6f51b\",\"size\":5321,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-create.md\",\"digest\":\"82dc3587acc5adcc0ace8798ed81c1ed25e5e49683d0142b407e7db4158ba9fa\",\"size\":3021,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-dimension.md\",\"digest\":\"e6b59217bb3969d9f4970ca0fcafaa6f94072d2b78ebd3e489f4ee7eb7aafe18\",\"size\":2369,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-dropdown.md\",\"digest\":\"0ba323acbf046ba64695942ed55241ee398c58cc120f6d53cd22ca485d174fa8\",\"size\":2011,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-filter-view-condition.md\",\"digest\":\"3b1a3431125d7eb9459e13bec68e9acc7c5e14627dbaff02105dd01b85550e68\",\"size\":940,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-filter-view.md\",\"digest\":\"7bddc0c1b49c9ddf9e6718a85432f04239642cf31a6008de06e02f67318990b7\",\"size\":844,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-delete-float-image.md\",\"digest\":\"e6e66e583552ab0b7e926b81b5f0b6f8c939c26ea01659bb2d6ae6a994cda193\",\"size\":1430,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-export.md\",\"digest\":\"466f59db8922f0dc7e89e84a36f2bf7527fd26d5a0d7063f832f9fc372cab002\",\"size\":2269,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-find.md\",\"digest\":\"d4e1864c082f78ee2e8808346bb19b81915cfd2c40f0110dba8d0e510a5b334a\",\"size\":3521,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-formula.md\",\"digest\":\"576bff8190cb9c86da25432a7fbed3fc59f238c2433ad4344d69e1fef145cba0\",\"size\":4233,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-get-dropdown.md\",\"digest\":\"da68446a6b43bde30fd444730cbc6ea7f0763a962130590d28e0df9bb20df57b\",\"size\":1740,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-get-filter-view-condition.md\",\"digest\":\"48987068ede198367c92dd6c2f62da7c1d8c32bdff38b48d16bd181ba001c1a4\",\"size\":947,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-get-filter-view.md\",\"digest\":\"e1cb58bd4bd03f2d7e76933177af5ba7e5a0080b6363903cc80fd1b49c40c2dd\",\"size\":807,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-get-float-image.md\",\"digest\":\"003f0450fc6773e2958d9d170c65124d6dbe0d2f8839bd9c9d1445dacbfae557\",\"size\":1732,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-info.md\",\"digest\":\"84867505a9f3802e313e5c32b05f3245f0b1e4a739fb7c27287a3c8b018afa97\",\"size\":1600,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-insert-dimension.md\",\"digest\":\"ae927c95deeab3d72e98492d13528c9ae554d10800bde2110cb50044bf0b3aab\",\"size\":2680,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-list-filter-view-conditions.md\",\"digest\":\"9674249bd0f26877bdfbed8fd647d25bdb0eb9d931843718d42f6faf922c8388\",\"size\":893,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-list-filter-views.md\",\"digest\":\"e04004bad2b4169e6fdf0b3505194b28dc736cf2232067613d9197a8f0a2b084\",\"size\":806,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-list-float-images.md\",\"digest\":\"a3c48c8c880528e6121cfd4229e5886007a7e51b3a9c109010122d2e2de46040\",\"size\":1602,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-media-upload.md\",\"digest\":\"26605e1e76220021f35924f5ea0988563d6d9ecba6fda65de9feeb2890e9db50\",\"size\":3575,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-merge-cells.md\",\"digest\":\"a46a7c94fe94f43eb830b63409bd9542e42f3db8605e02a898d6145c0c8941c1\",\"size\":2199,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-move-dimension.md\",\"digest\":\"cd42d7b4ae679bf968d2fc37956154e45cbd4a4ae03082453156c70352d1857d\",\"size\":2341,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-read.md\",\"digest\":\"8b5db888a721ae96f31e7da2b0537dc21796260b39072ac8c422e5f6b4fbb441\",\"size\":2917,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-replace.md\",\"digest\":\"cc34e86d9ba12e7575db3fd0c5006f9318cb15070e98cfab8a332dff1eeb763e\",\"size\":2854,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-set-dropdown.md\",\"digest\":\"6b8246c109e3c537ec17f3161f633cef512598422e4973e18ba3f5342975a802\",\"size\":3264,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-set-style.md\",\"digest\":\"f095783161e36676f85b46b0c31308b66a80db6dfefe65c4c8c54242b2bd6898\",\"size\":3764,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-unmerge-cells.md\",\"digest\":\"5ca41afbd297cb4db20a0b50c0908c3e890b0b57801cc33f82ea988d2536e106\",\"size\":1857,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-dimension.md\",\"digest\":\"83cffdaf374904e588ded164aaa1de698fc18b18615d416631e9781173a829a9\",\"size\":2838,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-dropdown.md\",\"digest\":\"bb5c8e6b444065f4a3555718c3052dd2728ca9061fa069fe9545661a0a347a2e\",\"size\":2444,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-filter-view-condition.md\",\"digest\":\"b61b8729a9601168f58effef09f2027c5a564575019fb748da42b31d6939cc16\",\"size\":1109,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-filter-view.md\",\"digest\":\"3e735566fb0fa1b3137678d72ff37d1d7904990544d2ccdd59c82079ab340204\",\"size\":865,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-update-float-image.md\",\"digest\":\"f9a3dee2aa3d9091685d5f6059758f7f78b45088dd1e05b1d4b8e99a8c332341\",\"size\":3689,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-write-image.md\",\"digest\":\"e2b97d27369e50b4d7d609eb22232672a52de37e1baf8baaf5318f396ea9ae12\",\"size\":3000,\"mode\":420},{\"path\":\"references/lark-sheets/lark-sheets-write.md\",\"digest\":\"d4e6e2136c3559100b4ee010abad78a5101db62010f0ba6f656ad310dfbeccfa\",\"size\":2802,\"mode\":420},{\"path\":\"references/lark-skill-maker.md\",\"digest\":\"13cb0c86eeb0747be0cdd5de05bb15cd2a29c81cc56c3aa9a9211dd4de82d3eb\",\"size\":2451,\"mode\":420},{\"path\":\"references/lark-slides.md\",\"digest\":\"ac5710259afe91d7fd0a538a605d45dd01362d66bd1ab5ab7f99dfa8d6a742ee\",\"size\":31154,\"mode\":420},{\"path\":\"references/lark-slides/examples.md\",\"digest\":\"3e465212249d8f7ec45460e009d3008aa3995824a82cc0840d81f37c9b333ff3\",\"size\":8657,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-create.md\",\"digest\":\"310fb9c269875b43ea5e4bfa29b9822656fb1a68968a2e43f182770889379508\",\"size\":7797,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-edit-workflows.md\",\"digest\":\"657ee5e2560f92a917d5ded2bf27709d62c37ed7e5e3473210d51ec520da9143\",\"size\":9054,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-media-upload.md\",\"digest\":\"156ef086de0c7cc3f41b94151bfe8a0c6b3871a0ce91b71646641abdcd4fda2c\",\"size\":6899,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-replace-slide.md\",\"digest\":\"5882afe329e99df2286950ee813a94c10dece90b07a99288686703587cb54094\",\"size\":15837,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentation-slide-create.md\",\"digest\":\"67af18aa008b7b31676d336000a6e1d1974a3158cbfad59d20f22293e4a6d1e4\",\"size\":9145,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentation-slide-delete.md\",\"digest\":\"aa6b82f30d476cb171e0ba7f98c60988f10936cbb602c110d8a902baa22abe1e\",\"size\":4594,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentation-slide-get.md\",\"digest\":\"3ff08d9def7e6ef69f5fd5a8a417b14700203b289b71821ac575e4cd07c026cc\",\"size\":3985,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentation-slide-replace.md\",\"digest\":\"9de02710406b7eb3c898fe3b715fb8334b3b9860217f19a0e78174f04cebb743\",\"size\":8735,\"mode\":420},{\"path\":\"references/lark-slides/lark-slides-xml-presentations-get.md\",\"digest\":\"67e99dd7601ba7de292448af1f29e2bdbb26829cb969797e57568660ece0a5b9\",\"size\":3425,\"mode\":420},{\"path\":\"references/lark-slides/slide-templates.md\",\"digest\":\"73c50b89cf5c7b2ca5408161df75d356fb62059d720220b9352d676e7a875e1a\",\"size\":10182,\"mode\":420},{\"path\":\"references/lark-slides/slides_demo.xml\",\"digest\":\"161034420a40441ee49cddeb04bf4c9381617651827be97fd959d054afc94d7e\",\"size\":12898,\"mode\":420},{\"path\":\"references/lark-slides/slides_xml_schema_definition.xml\",\"digest\":\"b2d8eaa71acb58786957a951c8110488a25d10b82ec32caf7e23aaa341cc0a32\",\"size\":173099,\"mode\":420},{\"path\":\"references/lark-slides/supported-fonts.md\",\"digest\":\"0ff1f9503744762c00d427310db1eed04ad556a5f0e18d7d1c2fd5c83c5bc9d6\",\"size\":2143,\"mode\":420},{\"path\":\"references/lark-slides/xml-format-guide.md\",\"digest\":\"043cebac11862bc4e6181f36876bce00faf88b36b7b1c3e220485e5844779b03\",\"size\":10990,\"mode\":420},{\"path\":\"references/lark-slides/xml-schema-quick-ref.md\",\"digest\":\"6fe9b8baa6270ca0abe2d57cf3cf2958be3e69bc0c6518c32616322772c821dd\",\"size\":7434,\"mode\":420},{\"path\":\"references/lark-task.md\",\"digest\":\"06b6799eac93a974a1eb3fb7c35d89a5a90e93c158c300008d19d4dec5fc8fb8\",\"size\":9925,\"mode\":420},{\"path\":\"references/lark-task/lark-task-assign.md\",\"digest\":\"5345d4b8784f7fee9270865ae52d4f83187032c8971823dd6970667f1edcc3c1\",\"size\":1684,\"mode\":420},{\"path\":\"references/lark-task/lark-task-comment.md\",\"digest\":\"3ba20bd078272aee771e850219324ab5c0bdf6e77e10e592e36b6dc2b77f358c\",\"size\":1317,\"mode\":420},{\"path\":\"references/lark-task/lark-task-complete.md\",\"digest\":\"cef972c53f76c4a4b312732f7ea5532692d9e194367339b5f6a264c2441d9969\",\"size\":1031,\"mode\":420},{\"path\":\"references/lark-task/lark-task-create.md\",\"digest\":\"b7aae12593c7972ecf4e6cafb5e7117cd9a81d5e77c8ffa3ba94e2827f639a54\",\"size\":3564,\"mode\":420},{\"path\":\"references/lark-task/lark-task-followers.md\",\"digest\":\"820bfa6ebd3fe1192eafa901fc3b29d14cef1f3ffb63a2ab56d691fe4fb1d51d\",\"size\":1549,\"mode\":420},{\"path\":\"references/lark-task/lark-task-get-my-tasks.md\",\"digest\":\"784c71b630cadbfdabf0103263109f9e326de7c124b7379e3accaee7e3d2d8fb\",\"size\":3863,\"mode\":420},{\"path\":\"references/lark-task/lark-task-get-related-tasks.md\",\"digest\":\"0fb9365dfc5875b0ad2b63bb9b7d07f7d7c695798b8587c257d97170c90219bb\",\"size\":3521,\"mode\":420},{\"path\":\"references/lark-task/lark-task-reminder.md\",\"digest\":\"ec9b79fa0324fcdc82e30a5b1aaf9d273fe24d8a5469f4724da16e1f28eaa575\",\"size\":1921,\"mode\":420},{\"path\":\"references/lark-task/lark-task-reopen.md\",\"digest\":\"2b350a0fab8e4bb3664264082529b31f45e2e3b2b3e190b6dc7ab522081fe8bb\",\"size\":1027,\"mode\":420},{\"path\":\"references/lark-task/lark-task-search.md\",\"digest\":\"912235f157b012e527b1d1f4b2a00fb0129cb5abe6e7c3543e17433ee84a63af\",\"size\":2196,\"mode\":420},{\"path\":\"references/lark-task/lark-task-set-ancestor.md\",\"digest\":\"c3a947d8394eb04891a636bcd3edb65550c08fb6ea38750af7436b80ae003471\",\"size\":1116,\"mode\":420},{\"path\":\"references/lark-task/lark-task-subscribe-event.md\",\"digest\":\"7ade9e0d5d225c93f53332a8b07ac8c160587227c7891e8fdbdde58d9b2fedc7\",\"size\":2816,\"mode\":420},{\"path\":\"references/lark-task/lark-task-tasklist-create.md\",\"digest\":\"548343a71b17d3da7ea8c53078fee6904819ef7c1450ca878859aea90523e9fb\",\"size\":1578,\"mode\":420},{\"path\":\"references/lark-task/lark-task-tasklist-members.md\",\"digest\":\"f8b6bc395ad67e46b02d72b1595845e78bfedd9d1aa13ef7723b1a35350ad0b1\",\"size\":1524,\"mode\":420},{\"path\":\"references/lark-task/lark-task-tasklist-search.md\",\"digest\":\"cf10c6bffb80391c01313e45b1c2121886e99d0bf88b32d126379cdc579fe83a\",\"size\":1891,\"mode\":420},{\"path\":\"references/lark-task/lark-task-tasklist-task-add.md\",\"digest\":\"a969be95a6f6ae2c32f575bff5b2ef6734fd18ae21fb4fd7ae8909a1b23ca410\",\"size\":2126,\"mode\":420},{\"path\":\"references/lark-task/lark-task-update.md\",\"digest\":\"751d88b63248b363cbc722d977e4ec5a2ac8a9b0546f7424d7f9398516ca7a75\",\"size\":2534,\"mode\":420},{\"path\":\"references/lark-vc.md\",\"digest\":\"bcff228438452ed5ab1432f5ee1a3ae90e76f018dcc635cecc6a24879d739cd1\",\"size\":10063,\"mode\":420},{\"path\":\"references/lark-vc/lark-vc-notes.md\",\"digest\":\"0c56efca3a2b240ee5a2d629f9341ec67d87af600b68d6be4bdaf0a2b7bef6ec\",\"size\":9945,\"mode\":420},{\"path\":\"references/lark-vc/lark-vc-recording.md\",\"digest\":\"6931c153247eda34ba726ff4657579bae1c2051ee8d003e89881a66eda1c1147\",\"size\":6728,\"mode\":420},{\"path\":\"references/lark-vc/lark-vc-search.md\",\"digest\":\"f8c81f0ed85a1be78035be719e6de022aa44d0340069672e3ed38ba0b8eabb93\",\"size\":7745,\"mode\":420},{\"path\":\"references/lark-whiteboard.md\",\"digest\":\"dd165ae69ce1d82c74be8579ef658353ddce11bf14fc9d5406c46d3ff89ca9ff\",\"size\":9488,\"mode\":420},{\"path\":\"references/lark-whiteboard/connectors.md\",\"digest\":\"eed35832652927557318814d42f613afbf8d7d71e8a64c942084fda5d504949a\",\"size\":5734,\"mode\":420},{\"path\":\"references/lark-whiteboard/content.md\",\"digest\":\"a4c279552fc9f21f5049a1a07d42828594a8b1ca529d5afa8cbf7529d40e57db\",\"size\":1983,\"mode\":420},{\"path\":\"references/lark-whiteboard/image.md\",\"digest\":\"fe1ecd0048efecb5bca013322e79228cceca1cf3ec541def6af5fd2b75aa5c2c\",\"size\":4563,\"mode\":420},{\"path\":\"references/lark-whiteboard/lark-whiteboard-query.md\",\"digest\":\"289cbc32bb17b351956e385aa98f40ed2005bbc939dae5a25025b750cd3a7016\",\"size\":2268,\"mode\":420},{\"path\":\"references/lark-whiteboard/lark-whiteboard-update.md\",\"digest\":\"340b922e777305f3abeeaa58ec6c637f096a6c849c7b4ed5efb10b13b19287a2\",\"size\":3729,\"mode\":420},{\"path\":\"references/lark-whiteboard/layout.md\",\"digest\":\"cd2916a1d8bc00cba39ca8eed8ab3bf205cff6354517008e4c6c94a7d6b6eb51\",\"size\":20166,\"mode\":420},{\"path\":\"references/lark-whiteboard/schema.md\",\"digest\":\"02ec326b7e312780b7d72432b0d99cfd8b0e3b7d3d1529007462786ff965cc8d\",\"size\":15575,\"mode\":420},{\"path\":\"references/lark-whiteboard/style.md\",\"digest\":\"b1e17fc2990ff46af63c7b040a0a32a6c0ffca837166e9e9ab631f78415676e5\",\"size\":12176,\"mode\":420},{\"path\":\"references/lark-whiteboard/typography.md\",\"digest\":\"b78a0e16754be6f9e0d2239034214ed3264c7076dfe05a45e5e0cc81e74760c1\",\"size\":2535,\"mode\":420},{\"path\":\"references/lark-wiki.md\",\"digest\":\"97830f971f77bdfac194ca85f7b304471d1f313bd564ce09cda5f8b8dff4100e\",\"size\":7091,\"mode\":420},{\"path\":\"references/lark-wiki/lark-wiki-delete-space.md\",\"digest\":\"b4217cb81641d49e87dccaf93b137be86c7412690949eaf2f32ab06a2fee65e2\",\"size\":8680,\"mode\":420},{\"path\":\"references/lark-wiki/lark-wiki-move.md\",\"digest\":\"dcc93217ccd34c76e750d3cf3d5b4af06aa0eb78bfbef1f6e8576fcd060d92f2\",\"size\":10270,\"mode\":420},{\"path\":\"references/lark-wiki/lark-wiki-node-create.md\",\"digest\":\"ded5acfb262525d6a8aadc80e293fc37acdf799c3528a6bc6c1ddb509ff189cf\",\"size\":7176,\"mode\":420},{\"path\":\"references/lark-workflow-meeting-summary.md\",\"digest\":\"b7adbbd53ace2f639a01a8683733665c5b3311dbadd150b6f7b37e9780354b9c\",\"size\":4125,\"mode\":420},{\"path\":\"references/lark-workflow-standup-report.md\",\"digest\":\"668337ff574177747d7e255586d975568e793d268e2fb3dd330036b31742f202\",\"size\":4456,\"mode\":420}],\"digest\":\"1e0ed8834390381e680cfcce513003d52c193938fe07e188a4d8557cbb3c5039\"},{\"ref\":\"builtin:python-script\",\"api_id\":\"builtin-python-script\",\"name\":\"python-script\",\"description\":\"Create robust Python automation with full logging and safety checks. Use when tasks need complex data processing, authenticated API work, conditional file operations, or error handling beyond simple shell commands.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{},\"root\":\"system/python-script\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"f900fc35795fc1018fe086a5cd2b9d7682511960274b5700b2c4774c676266fb\",\"size\":1299,\"mode\":420},{\"path\":\"references/script-template.md\",\"digest\":\"7e681d0db5c956347d7d5e7b0380725beaf536e32375fcf79b8dac4e70a8ff02\",\"size\":1060,\"mode\":420}],\"digest\":\"7837df183e21e2b0fd9fc1e37943b796db632a40ae7f01243566b6a8d9c812b3\"},{\"ref\":\"builtin:recally\",\"api_id\":\"builtin-recally\",\"name\":\"recally\",\"description\":\"Reading assistant for saving, organizing, and recalling web content. Use when the user says \\\"save this article\\\", \\\"read this link\\\", \\\"summarize this\\\", \\\"check my feeds\\\", \\\"add to my library\\\", or asks about previously saved content. Handles articles, tweets, YouTube videos, GitHub repos, PDFs, and RSS feeds. Articles are stored with their metadata and indexed for fast search.\\n\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"author\":\"CherryHQ/stella\",\"owner_plugin\":\"system/recally\",\"version\":\"1.0\"}},\"root\":\"system/recally\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"a42fb637f661a46f38bae3a4a9282830104dfd53e4bf08a9a5d087f853f20f7d\",\"size\":7195,\"mode\":420},{\"path\":\"references/rss-workflow.md\",\"digest\":\"5e84e29d4e29686c62c0034402aa6085eff17286a8e4c749aa90ab8820a12ed6\",\"size\":1686,\"mode\":420},{\"path\":\"references/save-workflow.md\",\"digest\":\"d8f38e42b647a3b654b0a31932da2034a9125ae4ffa6a12d535ba769ec1a3ee1\",\"size\":4911,\"mode\":420},{\"path\":\"references/twitter-workflow.md\",\"digest\":\"cfc415312a5a0bcc73c23e93cde52ee5002e846fad41798261b532650910801b\",\"size\":2265,\"mode\":420},{\"path\":\"references/website-workflow.md\",\"digest\":\"2789838d5fc5d6b33322b34fe8a3c4508851ada7a99593ae2783c5f1587713aa\",\"size\":2682,\"mode\":420}],\"digest\":\"ad2f4a3c9d6addb28c38676da6df82676484b351cce09dd5cf746817c610f3f5\"},{\"ref\":\"builtin:scheduler\",\"api_id\":\"builtin-scheduler\",\"name\":\"scheduler\",\"description\":\"Manage scheduled jobs. Use when the user wants to create, list, inspect, update, pause, resume, or remove recurring or one-time scheduled tasks. Handles cron schedules, interval-based (every), one-time (at) jobs, platform job templates, and scheduled workflow runs.\\n\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"author\":\"CherryHQ/stella\",\"owner_plugin\":\"system/scheduler\",\"version\":\"1.0\"}},\"root\":\"system/scheduler\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"972cf6aaa251cfea29732b9d3d234ecf6f2fed8b787e59145e4f5861dce719f0\",\"size\":4439,\"mode\":420}],\"digest\":\"199deb7e2ccbe1eb4a3323c61544a3e3450cf6d5e1689410c360356850d73de0\"},{\"ref\":\"builtin:skill-creator\",\"api_id\":\"builtin-skill-creator\",\"name\":\"skill-creator\",\"description\":\"Create new skills, modify and improve existing skills, and measure skill performance. Use when users want to create a skill from scratch, edit, or optimize an existing skill, run evals to test a skill, benchmark skill performance with variance analysis, or optimize a skill's description for better triggering accuracy.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{},\"root\":\"system/skill-creator\",\"files\":[{\"path\":\"LICENSE.txt\",\"digest\":\"58d1e17ffe5109a7ae296caafcadfdbe6a7d176f0bc4ab01e12a689b0499d8bd\",\"size\":11357,\"mode\":420},{\"path\":\"SKILL.md\",\"digest\":\"baa75a289d3573958f48d6c0494833875c1240f64c76b256a0c1803ee14bbb52\",\"size\":33282,\"mode\":420},{\"path\":\"agents/analyzer.md\",\"digest\":\"28ff495369f27d9e2c85c133c85b596e9cfb3b672252e95b8f5ca56cc6e012f5\",\"size\":10561,\"mode\":420},{\"path\":\"agents/comparator.md\",\"digest\":\"58a5e00788c709dfba4090292e5f351cb40e3fd70975feecc1a4257d2cfcccfa\",\"size\":7569,\"mode\":420},{\"path\":\"agents/grader.md\",\"digest\":\"7a9ba21f96d846eacb438b5d8d97ee6d2d74589395cef788b29f5140b72e9322\",\"size\":9064,\"mode\":420},{\"path\":\"assets/eval_review.html\",\"digest\":\"1f9ba52cfd4f9cf528a2f26acfe097f15c92f048b947f28e84555cd564a9db2f\",\"size\":7184,\"mode\":420},{\"path\":\"eval-viewer/generate_review.py\",\"digest\":\"fc9d1b9243fe5ab6012ebd579bd76d0035de1b79fd3b969de114defab26478fb\",\"size\":16365,\"mode\":420},{\"path\":\"eval-viewer/viewer.html\",\"digest\":\"8ba4b172fdc74f009d087a7af06b7c7c7aefb19da0bf9b14d96892c72aa5c315\",\"size\":45753,\"mode\":420},{\"path\":\"references/schemas.md\",\"digest\":\"467025efdc41dd3e4171d992a9f28adc2a9cba27c0bb45478598737553c9f86a\",\"size\":12172,\"mode\":420},{\"path\":\"scripts/__init__.py\",\"digest\":\"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855\",\"size\":0,\"mode\":420},{\"path\":\"scripts/aggregate_benchmark.py\",\"digest\":\"123ef128ea5ccc01a4b1ac212ef5567f21e9c13d3d240609780beeb3200c49aa\",\"size\":14386,\"mode\":493},{\"path\":\"scripts/generate_report.py\",\"digest\":\"13df7118a3c50c83c4c3250a606d5f2b20b25a3d44cbc392b3d669ec75281453\",\"size\":12847,\"mode\":493},{\"path\":\"scripts/improve_description.py\",\"digest\":\"87d864570220b699fac52da309d2d6efdb060647bfebc74f768128e646accf80\",\"size\":11116,\"mode\":493},{\"path\":\"scripts/package_skill.py\",\"digest\":\"1a33059b0db1ef73375d46d513e5ea81369d2e8838c970597b0d52ddef8d1c0f\",\"size\":4234,\"mode\":493},{\"path\":\"scripts/quick_validate.py\",\"digest\":\"67cf5703402013936c8fb75ad6a1afecd8841d45cc5e606b634eb05825fde365\",\"size\":3972,\"mode\":493},{\"path\":\"scripts/run_eval.py\",\"digest\":\"43e3b8f80dbf69c343967ba77e268fae991d9fa3ed68b32a0ff02532cd48657f\",\"size\":11464,\"mode\":493},{\"path\":\"scripts/run_loop.py\",\"digest\":\"7bd6f674203168520517eec94c55f493c0d154339b061b4d7c0f0dad187d0f21\",\"size\":13605,\"mode\":493},{\"path\":\"scripts/utils.py\",\"digest\":\"3af8ae62c40c73ab712207436a0d9a981e845f25c5a7040229eb189cc8e45bb1\",\"size\":1661,\"mode\":420}],\"digest\":\"414822d4852b4511d8be6a799f58fc10712f370a83941844d8097dfb863a03ee\"},{\"ref\":\"builtin:stella\",\"api_id\":\"builtin-stella\",\"name\":\"stella\",\"description\":\"Self-knowledge about stella, the self-hosted AI assistant. Use when the user asks about stella itself: configuration, setup, onboarding, providers, models, agents, channels (Telegram/Discord/QQ/Feishu/DingTalk/WeChat), webhooks, memory system (LCM), scheduled jobs, reusable workflows, goals (objectives that converge through acceptance), workers/decomposition/dependencies, skills, plugins, session compaction, notifications, self-update, multi-agent, multi-user, or general \\\"how does stella work\\\" / \\\"help me get started\\\" questions. Also triggers on \\\"change my model\\\", \\\"set up telegram\\\", \\\"set up discord\\\", \\\"set up dingtalk\\\", \\\"set up wechat\\\", \\\"set up webhook\\\", \\\"configure provider\\\", \\\"update stella\\\", \\\"what can you do\\\", \\\"how do I install skills\\\", \\\"stella onboard\\\", \\\"switch agent\\\". Also triggers when the user wants to report a bug or file a GitHub issue about stella: \\\"report this bug\\\", \\\"create an issue for this\\\", \\\"报告这个 issue\\\", \\\"帮我建个 issue\\\". Also triggers on goal-model work — goal status, \\\"why is this blocked\\\", \\\"decompose this\\\", \\\"review/accept this\\\", \\\"为什么卡住了\\\", \\\"拆解成子目标\\\", \\\"验收\\\", or being dispatched as a goal worker: read references/goals.md. Merely creating a goal does not — write a clear intent and call the goal tool.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{},\"root\":\"system/stella\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"7a4d0c296004020d56f73a4b652d99026632341a0c2894e0d21e1cdd2133073a\",\"size\":28088,\"mode\":420},{\"path\":\"references/channels.md\",\"digest\":\"74137d86ca1d23e6d279ed8512d29649b9b0720994dfd69bf9cdb27df5167cd0\",\"size\":12135,\"mode\":420},{\"path\":\"references/configuration.md\",\"digest\":\"a4a1a31e7f7bfdb1dccab4f2c4bc531fc8cda3fdc568af34bd2891cb5f6cfb67\",\"size\":29273,\"mode\":420},{\"path\":\"references/goals.md\",\"digest\":\"056072761f3433b4005bf7aa912923c2d9883e28b90eb22af056ef990785654a\",\"size\":8885,\"mode\":420},{\"path\":\"references/models.md\",\"digest\":\"0734e738887b92e2dd8949f1132ce579bfd45ed53b747aef82994ecee88f6800\",\"size\":3001,\"mode\":420},{\"path\":\"references/report-issue.md\",\"digest\":\"dcc465955a7e95dadb3cf6bc8948f6c3c75bee24781719c19bebb2cd6a05be51\",\"size\":3347,\"mode\":420},{\"path\":\"references/update.md\",\"digest\":\"b3df654900385845fac46af4b01fafe685766af6336a749cf6cb2e96e87b660f\",\"size\":3700,\"mode\":420},{\"path\":\"references/webhooks.md\",\"digest\":\"f21d41006e7a1d5c4723f93535e416756751f475322e6ee745d71188bcdddbc3\",\"size\":1906,\"mode\":420}],\"digest\":\"8940425ef48f72a2b2a64591f6f3224d5003de0ed83abd8be050baa928b62b0c\"},{\"ref\":\"builtin:web\",\"api_id\":\"builtin-web\",\"name\":\"web\",\"description\":\"Read the public web: search for sources, read one page as Markdown, and run site scripts for structured records (X/Twitter via FxEmbed, Exa search, GitHub, Hacker News, Reddit, Wikipedia, Bilibili bundled; a catalog of more installable with one command). Use for any task that needs something from a public website: a topic without a URL, a URL to read, a site's records (a tweet, a timeline, a repo's stats, a front page, a ranking), or a page that only renders with JavaScript.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"metadata\":{\"author\":\"vaayne/tap\",\"version\":\"2.0.0\"}},\"root\":\"system/web\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"23fd58ef408a2d23d625ba045b054f5c38baa377af7454112ad4c14a0241ab10\",\"size\":11306,\"mode\":420},{\"path\":\"scripts/lib/fetch.ts\",\"digest\":\"db9d52ecfef2e3795822035a7fade8bc3b727ec5c12704acb93996481dcc1f87\",\"size\":10275,\"mode\":420},{\"path\":\"scripts/lib/http.ts\",\"digest\":\"7caa499ea938e9a25c0bddcb4ec1f44d8e87451b9dde1ffa1e4e1994869bfd9a\",\"size\":2635,\"mode\":420},{\"path\":\"scripts/lib/search.ts\",\"digest\":\"af75b7bde35573ece9fb02a9fc3e9e5e00ae05a8fabef9978ac59e5ae032f295\",\"size\":10103,\"mode\":420},{\"path\":\"scripts/lib/site.ts\",\"digest\":\"2138b9f6a50a15a8182843449c12543c556769b0db99912e090b0ac1d4963212\",\"size\":18296,\"mode\":420},{\"path\":\"scripts/package.json\",\"digest\":\"72163e64f1d71a6d466645d5490a902e42a75d96118f86b3b2f073a55ff8b8a3\",\"size\":149,\"mode\":420},{\"path\":\"scripts/web.ts\",\"digest\":\"423cb4aea2e512e0af56f8cf0846b0be8fd188a779a814aa41597bad17ffc3c0\",\"size\":5260,\"mode\":420},{\"path\":\"sites/bilibili/video.js\",\"digest\":\"f36fefce96dabc9ee0b1c42f24f52cf09f7d468356985ab1ddc7e910236f49c4\",\"size\":2321,\"mode\":420},{\"path\":\"sites/exa/search.js\",\"digest\":\"e08ca9ffc6311fd96bd35bfe7d6567860c20a668e60d6f3d4e0da2335dec50a1\",\"size\":3396,\"mode\":420},{\"path\":\"sites/github/repo.js\",\"digest\":\"16fbd3d0bf103d63d087afe6dd50ddeea7b86e8754ace7ae5846762c633ddb59\",\"size\":2234,\"mode\":420},{\"path\":\"sites/hackernews/top.js\",\"digest\":\"68fb21108eda4e7d4cd012c002b08c3b5b764a7ca66289512605bb7d721ac49e\",\"size\":3799,\"mode\":420},{\"path\":\"sites/reddit/search.js\",\"digest\":\"8cfdd1688706fbbc570315e48974b589b3c928975106340c95d3faf7df4944f9\",\"size\":4243,\"mode\":420},{\"path\":\"sites/twitter/fxembed-profile-articles.js\",\"digest\":\"89a0c518360b845def92609c76b474d96c2066eae1477557c4f09e7f8e744eb5\",\"size\":1340,\"mode\":420},{\"path\":\"sites/twitter/fxembed-profile-statuses.js\",\"digest\":\"da3fcfb2ff1b972041e75497d780b40b6cbd5668d5a25ecb03521c33aaf1576b\",\"size\":1973,\"mode\":420},{\"path\":\"sites/twitter/fxembed-status.js\",\"digest\":\"64e030d0dd4df7e1b176b16217d5a14300045343cb679c4bc9b830c36b659efb\",\"size\":3529,\"mode\":420},{\"path\":\"sites/wikipedia/summary.js\",\"digest\":\"ea54081ec7c031c075513f62e7a9e67f89448486c2b70334391f1851bb0f889c\",\"size\":891,\"mode\":420}],\"digest\":\"54a9421f2372542ed80e7aa10bfdd991a396a1ecf212bd674ba6e887e172ac0b\"},{\"ref\":\"builtin:xberg\",\"api_id\":\"builtin-xberg\",\"name\":\"xberg\",\"description\":\"Extract text, tables, metadata, and images from 97+ document formats (PDF, Office, images, HTML, email, archives, academic) using Xberg CLI.\",\"tags\":null,\"disable_model_invocation\":false,\"metadata\":{\"license\":\"MIT\",\"metadata\":{\"author\":\"xberg-io\",\"repository\":\"https://github.com/xberg-io/xberg\",\"version\":\"1.0.14\"}},\"root\":\"system/xberg\",\"files\":[{\"path\":\"SKILL.md\",\"digest\":\"110b3822dd43783022eb40a0aa7ff91e5a58eadc64ef7b6f95bb2a097f870c7e\",\"size\":4161,\"mode\":420},{\"path\":\"references/supported-formats.md\",\"digest\":\"8712ba5e292662a875b666574d81542d5527c4843e7d2bbba3e05b5092090983\",\"size\":17799,\"mode\":420}],\"digest\":\"f7e4268303b9e8f4633301aea3b3895399e3c61b239e4154439e211e96331b96\"}]}"
