+++ web/content/docs/changelog.mdx
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### ✨ Added
 
-- `web` skill and `tool/lightpanda` manifest plugin: every public-web need is one skill. `bun scripts/web.ts search "<query>"` finds sources through the configured providers (Firecrawl, Parallel, Tavily, Exa, Jina, SearXNG, Brave, Keenable, each enabled by its API-key variable, usually a vault secret) and falls back to Exa's anonymous endpoint; `bun scripts/web.ts fetch <url>` reads one page as Markdown with Defuddle, renders it with Lightpanda when the plain HTML has no body, then asks Jina Reader; `python3 scripts/site.py run <site/name> key=value` runs Tap-format site scripts through the Lightpanda nightly headless browser CLI. Nine scripts are bundled (X/Twitter posts and articles via FxEmbed, Exa search, GitHub, Hacker News, Reddit, Wikipedia, Bilibili); `site.py add <site/name | url | file>` installs any of the ~140 Tap catalog scripts, or your own, into `$XDG_CACHE_HOME/site-scripts/`, shared by all of a user's agents. Scripts need no login and no Chrome. The skill is on the default Agent template; add `web` to a custom template's `skills:` list to enable it.
+- `web` skill and `tool/lightpanda` manifest plugin: every public-web need is one skill, and the skill is a single bun program (`scripts/web.ts` plus its `scripts/lib/` modules), no Python. `bun scripts/web.ts search "<query>"` finds sources through the configured providers (Firecrawl, Parallel, Tavily, Exa, Jina, SearXNG, Brave, Keenable, each enabled by its API-key variable, usually a vault secret) and falls back to Exa's anonymous endpoint; `bun scripts/web.ts fetch <url>` reads one page as Markdown with Defuddle, renders it with Lightpanda when the plain HTML has no body, then asks Jina Reader; `bun scripts/web.ts site run <site/name> key=value` runs Tap-format site scripts through the Lightpanda nightly headless browser CLI. Nine scripts are bundled (X/Twitter posts and articles via FxEmbed, Exa search, GitHub, Hacker News, Reddit, Wikipedia, Bilibili); `web.ts site add <site/name | url | file>` installs any of the ~140 Tap catalog scripts, or your own, into `$XDG_CACHE_HOME/site-scripts/`, shared by all of a user's agents. Scripts need no login and no Chrome. The skill is on the default Agent template; add `web` to a custom template's `skills:` list to enable it.
 
 ### ⚠️ Breaking Changes
